### 2016-12-05 #126 @whedon
[PRE REVIEW]: Effective-Quadratures: Polynomials for Computational Engineering Studies
**Submitting author:** @psesh (<a href="http://orcid.org/0000-0002-7351-012X">Pranay Seshadri</a>)
**Repository:** <a href="https://github.com/Effective-Quadratures/Effective-Quadratures/blob/master/joss/paper.md" target ="_blank">https://github.com/Effective-Quadratures/Effective-Quadratures/blob/master/joss/paper.md</a>
**Version:** v3.0
