### 2016-12-05 #126 @arfon
@katyhuff - would you be able to edit this submission for JOSS?
