### 2016-12-05 #126 @katyhuff
Yes!

On Mon, Dec 5, 2016 at 6:52 AM, Arfon Smith <notifications@github.com>
wrote:

> @katyhuff <https://github.com/katyhuff> - would you be able to edit this
> submission for JOSS?
>
