### 2016-12-05 7f63226
Delete tke_inside.txt
