+++ ipython-notebooks/Vegetation_results/tke_inside.txt
@@ -1,16 +1,0 @@
-Case_no, Mean TKE inside(mts2/sec3), Mean_bottom_TKE inside: 
-1	, 0.982942027286	  0.0157066565149	 
-2	, 0.921938664576	  0.0152814566145	 
-3	, 1.00617955475	  0.0158136200045	 
-4	, 0.974924878894	  0.0162772074051	 
-5	, 0.974957862607	  0.0152599081594	 
-6	, 0.934422493103	  0.01568993954	 
-7	, 0.973911430065	  0.0164453320808	 
-8	, 1.01473064236	  0.0155030335447	 
-9	, 0.987672040492	  0.0160158048898	 
-10	, 0.999911373132	  0.0158579727413	 
-11	, 0.933584728739	  0.0158549165056	 
-12	, 0.954102576587	  0.0153885373966	 
-13	, 0.927377135023	  0.0158448447729	 
-14	, 0.967593456847	  0.0155461085033	 
-15	, 0.940013967219	  0.0157530721124	 
