### 2016-12-05 0a6fdce
Delete ke_output.txt
