+++ ipython-notebooks/Vegetation_results/ke_output.txt
@@ -1,16 +1,0 @@
-Case_no, Mean KE (mts2/sec2), Mean_Bottom_KE(mt2/sec2): 
-1	, 436.770095328	, 198.709490963	 
-2	, 524.746650218	, 321.344019678	 
-3	, 411.681338502	, 172.189065624	 
-4	, 445.17901124	, 195.967796404	 
-5	, 450.962042491	, 225.952392289	 
-6	, 498.159288373	, 267.81265041	 
-7	, 449.866057827	, 197.813005987	 
-8	, 392.910603292	, 165.45354121	 
-9	, 430.181869637	, 183.728742474	 
-10	, 416.961026324	, 176.471953008	 
-11	, 500.449092724	, 268.249207277	 
-12	, 475.765464167	, 248.429405886	 
-13	, 509.803436682	, 285.714071881	 
-14	, 456.077067392	, 222.452267595	 
-15	, 494.391117514	, 264.857478423	 
