### 2016-12-08 9a87e33
new function templates
