### 2016-12-09 f566e67
inverse CFD for beta distribution
