+++ ipython-notebooks/vegetation_12_dec/parameter_space.txt
@@ -1,0 +1,16 @@
+     Density        Height           Diameter        Thickness
+1  [ 1.44300000e+02, 2.35000000e-01, 6.00000000e-03, 6.00000000e-04],
+2  [ 6.21152934e+01, 1.74581460e-01, 2.90161332e-03, 6.00000000e-04],
+3  [ 2.26484707e+02, 2.95418540e-01, 6.00000000e-03, 2.90161332e-04],
+4  [ 1.44300000e+02, 1.74581460e-01, 9.09838668e-03, 9.09838668e-04],
+5  [ 1.44300000e+02, 2.95418540e-01, 2.90161332e-03, 9.09838668e-04],
+6  [ 6.21152934e+01, 2.95418540e-01, 6.00000000e-03, 2.90161332e-04],
+7  [ 2.26484707e+02, 1.74581460e-01, 6.00000000e-03, 2.90161332e-04],
+8  [ 1.44300000e+02, 2.95418540e-01, 9.09838668e-03, 9.09838668e-04],
+9  [ 1.44300000e+02, 2.35000000e-01, 9.09838668e-03, 2.90161332e-04],
+10 [ 2.26484707e+02, 2.35000000e-01, 6.00000000e-03, 9.09838668e-04],
+11 [ 1.44300000e+02, 2.35000000e-01, 2.90161332e-03, 2.90161332e-04],
+12 [ 6.21152934e+01, 2.35000000e-01, 6.00000000e-03, 9.09838668e-04],
+13 [ 6.21152934e+01, 1.74581460e-01, 6.00000000e-03, 2.90161332e-04],
+14 [ 6.21152934e+01, 2.35000000e-01, 9.09838668e-03, 6.00000000e-04],
+15 [ 1.44300000e+02, 1.74581460e-01, 2.90161332e-03, 9.09838668e-04]])
