+++ ipython-notebooks/vegetation_12_dec/dissip_output.txt
@@ -1,0 +1,16 @@
+Case_no, Mean wave dissipation, percent_change_dissip: 
+1	, 0.00288641956093	, 390.809208985	 
+2	, 0.000588094010481	, 0.0	 
+3	, 0.00406629056334	, 591.435466248	 
+4	, 0.0030033701336	, 410.695582692	 
+5	, 0.00238950553151	, 306.313529628	 
+6	, 0.00147861895071	, 151.425609573	 
+7	, 0.00351944094428	, 498.448697242	 
+8	, 0.0035562196531	, 504.702579812	 
+9	, 0.00332515724102	, 465.41253299	 
+10	, 0.00382217771319	, 549.926311962	 
+11	, 0.00217241293763	, 269.398922436	 
+12	, 0.00125030030237	, 112.602114643	 
+13	, 0.00102605317042	, 74.4709437828	 
+14	, 0.00160467600591	, 172.860457224	 
+15	, 0.0018514665637	, 214.824931169	 
