+++ ipython-notebooks/vegetation_12_dec/Grad_Water_level_output.txt
@@ -1,0 +1,16 @@
+Case_no, Grad_Water_level(mts.), pcnt_min_grad_ret, istore, jstore: 
+1	, 1.14128459245e-06	, 6.87268889974	, 55	, 59	 
+2	, 1.06789171696e-06	, 0.0		, 55	, 59	 
+3	, 1.11872330308e-06	, 4.75999441847	, 55	, 59	 
+4	, 1.12950801849e-06	, 5.76990162562	, 55	, 59	 
+5	, 1.13985966891e-06	, 6.73925556408	, 55	, 59	 
+6	, 1.08966603875e-06	, 2.039000907	, 55	, 59	 
+7	, 1.10349152237e-06	, 3.33365310821	, 55	, 59	 
+8	, 1.16420444101e-06	, 9.01895974325	, 55	, 59	 
+9	, 1.12222507596e-06	, 5.08790902114	, 55	, 59	 
+10	, 1.13915186375e-06	, 6.67297495291	, 55	, 59	 
+11	, 1.09617132694e-06	, 2.64817205051	, 55	, 59	 
+12	, 1.11574307084e-06	, 4.48091816089	, 55	, 59	 
+13	, 1.08011066914e-06	, 1.14421265611	, 55	, 59	 
+14	, 1.13489106297e-06	, 6.27398311589	, 55	, 59	 
+15	, 1.10959168524e-06	, 3.90488732296	, 55	, 59	 
