+++ ipython-notebooks/vegetation_12_dec/ke_output.txt
@@ -1,0 +1,16 @@
+Case_no,      Mean KE,         Mean_bot_KE, Mean_sur_KE,   pcnt_mean_KE, pcnt_mean_bot_KE, pcnt_mean_sur_KE: 
+1	, 436.770095328	, 3.36795747394	, 8.52404761484	, 11.1627153018	, 20.0998718487	, 7.78754367859	 
+2	, 524.746650218	, 5.44650880811	, 9.60948216137	, 33.5537004658	, 94.2200918322	, 21.5129859663	 
+3	, 411.681338502	, 2.91845873939	, 8.20174447807	, 4.77735521852	, 4.07094605801	, 3.71198415546	 
+4	, 445.17901124	, 3.32148807464	, 8.6242807928	, 13.3028753897	, 18.4427936507	, 9.05500352112	 
+5	, 450.962042491	, 3.82970156422	, 8.73271472785	, 14.7747194178	, 36.565461601	, 10.4261628621	 
+6	, 498.159288373	, 4.53919746457	, 9.31612589742	, 26.7869291894	, 61.865771171	, 17.8034629154	 
+7	, 449.866057827	, 3.35276281333	, 8.69939127787	, 14.4957794619	, 19.558036981	, 10.0047840779	 
+8	, 392.910603292	, 2.80429730865	, 7.9081935852	, 0.0	, 0.0	, 0.0	 
+9	, 430.181869637	, 3.11404648261	, 8.43951671019	, 9.48594057608	, 11.0455183551	, 6.71864085346	 
+10	, 416.961026324	, 2.99105005099	, 8.25650055602	, 6.12109289764	, 6.6595200788	, 4.40438093815	 
+11	, 500.449092724	, 4.54659673351	, 9.342665916	, 27.3697091734	, 62.129625824	, 18.1390644443	 
+12	, 475.765464167	, 4.21066789638	, 9.03702532693	, 21.0874586182	, 50.1505522752	, 14.2742047165	 
+13	, 509.803436682	, 4.84261138781	, 9.43323458749	, 29.7504909289	, 72.6853772913	, 19.2843155122	 
+14	, 456.077067392	, 3.77037741686	, 8.77886017759	, 16.0765486019	, 34.4499887809	, 11.0096772797	 
+15	, 494.391117514	, 4.48910980379	, 9.25803327151	, 25.8278889325	, 60.0796673712	, 17.0688751074	 
