+++ ipython-notebooks/vegetation_12_dec/TKE_inside.txt
@@ -1,0 +1,16 @@
+Case_no, Mean_TKE 		,  Mean_bot_KE 		, pcnt_mean_TKE , pcnt_Mean_bot_TKE 	: 
+1	, 0.982942027286	, 0.0157066565149	, 6.61685696176	, 2.92759530911	 
+2	, 0.921938664576	, 0.0152814566145	, 0.0	,         0.141209599129	 
+3	, 1.00617955475	        , 0.0158136200045	, 9.13736384125	, 3.62853982706	 
+4	, 0.974924878894	, 0.0162772074051	, 5.74726024126	, 6.66648340925	 
+5	, 0.974957862607	, 0.0152599081594	, 5.75083788851	, 0.0	 
+6	, 0.934422493103	, 0.01568993954		, 1.3540844968	, 2.81804697687	 
+7	, 0.973911430065	, 0.0164453320808	, 5.63733440046	, 7.7682244809	 
+8	, 1.01473064236		, 0.0155030335447	, 10.064875393	, 1.59322967595	 
+9	, 0.987672040492	, 0.0160158048898	, 7.12990770883	, 4.95348151852	 
+10	, 0.999911373132	, 0.0158579727413	, 8.45747244935	, 3.91918860628	 
+11	, 0.933584728739	, 0.0158549165056	, 1.26321463784	, 3.89916072889	 
+12	, 0.954102576587	, 0.0153885373966	, 3.48872579566	, 0.842922748325	 
+13	, 0.927377135023	, 0.0158448447729	, 0.589895039234, 3.83315946164	 
+14	, 0.967593456847	, 0.0155461085033	, 4.95204225889	, 1.87550502233	 
+15	, 0.940013967219	, 0.0157530721124	, 1.96057539811	, 3.2317622612	 
