+++ ipython-notebooks/vegetation_12_dec/Water_level_output.txt
@@ -1,0 +1,16 @@
+Case_no, Water_level(mts.), pcnt_min_wl_ret, istore, jstore: 
+1	, 0.00513709	, 11.4752826138	, 45	, 42	 
+2	, 0.00460827	, 0.0		, 45	, 42	 
+3	, 0.00533104	, 15.6840934372	, 45	, 42	 
+4	, 0.00509891	, 10.6468427452	, 45	, 42	 
+5	, 0.00503859	, 9.33789481853	, 45	, 42	 
+6	, 0.00487947	, 5.88509196265	, 45	, 42	 
+7	, 0.00508943	, 10.4411878832	, 45	, 42	 
+8	, 0.00539345	, 17.038311302	, 45	, 42	 
+9	, 0.00520864	, 13.0280414931	, 45	, 42	 
+10	, 0.00525871	, 14.1145200093	, 45	, 42	 
+11	, 0.00483891	, 5.00491502186	, 45	, 42	 
+12	, 0.00490952	, 6.53697907235	, 45	, 42	 
+13	, 0.00475609	, 3.20769847634	, 45	, 42	 
+14	, 0.00502881	, 9.12577282252	, 45	, 42	 
+15	, 0.00479034	, 3.9507721758	, 45	, 42	 
