### 2016-12-14 f447b32
added output for indices
