+++ ipython-notebooks/vegetation_12_dec/ke_diff_top_botout.txt
@@ -1,0 +1,16 @@
+Case_no, mean_ke_diff_topbot, pcnt_mean_ke_diff_topbot: 
+1	, 5.15609014089	, 23.8559487019	 
+2	, 4.16297335326	, 0.0	 
+3	, 5.28328573868	, 26.911351343	 
+4	, 5.30279271816	, 27.3799342002	 
+5	, 4.90301316362	, 17.7767126417	 
+6	, 4.77692843285	, 14.7479944618	 
+7	, 5.34662846453	, 28.4329254796	 
+8	, 5.10389627655	, 22.6021846272	 
+9	, 5.32547022758	, 27.9246772839	 
+10	, 5.26545050503	, 26.4829259813	 
+11	, 4.7960691825	, 15.207780005	 
+12	, 4.82635743056	, 15.9353428668	 
+13	, 4.59062319968	, 10.2727019878	 
+14	, 5.00848276072	, 20.3102286685	 
+15	, 4.76892346773	, 14.5557048543	 
