+++ effective_quadratures/qr.py
@@ -13,8 +13,6 @@ from utils import error_function
 # 6. The 'Practical QR' algorithm
 # 7. Randomized QR 
 #****************************************************************************
-<<<<<<< Updated upstream
-
 # SVD via QR!
 def tsvd(A):
     return 0
@@ -24,7 +22,6 @@ def qrblocks(A, n):
     # n is the number of blocks!
     return 0
     
-=======
 def solveCLSQ(A,b,C,d):
     """
     Solves the direct, constraint least squares problem ||Ax-b||_2 subject to Cx=d using 
@@ -56,23 +53,44 @@ def solveCLSQ(A,b,C,d):
     elif k != s:
         error_function('ERROR: mismatch in sizes of C and d') 
 
-    Q , R = qr_Householder(C, 1) # Thin QR factorization on C'
-    R = R[0:n, 0:n]
+    notused , R = qr(C.T, thin=1)#qr_Householder(C.T, 1) # Thin QR factorization on C'
+    Q , notused = qr(C.T, thin=0)#qr_Householder(C.T)
+
+    print '***Debugging from qr.py****'
+    print Q
+    print R
+    print '~~~~~~~~~~~~~~~'
+    # up to here is OK! - Pranay Dec 16th 2016
+
     u = np.linalg.inv(R.T) * d
+    print '---- Check u ----'
+    print u
+    print '####################'
     A_hat = A * Q
     z, w = A_hat.shape
-    
+
     # Now split A
     Ahat_1 = A_hat[:, 0:len(u)]
     Ahat_2 = A_hat[:, len(u) : w]
     r = b - Ahat_1 * u
+    print '***Check Ahat_1 and Ahat_2 ****'
+    print Ahat_1
+    print Ahat_2
+    print '~'
+    print A_hat
+    print r
+    print '$$$$$$$$$$$$$'
+    
     
     # Solve the least squares problem
     v = solveLSQ(Ahat_2, r)
     x = Q * np.vstack([u, v]) 
+    print '** Final steps ***'
+    print v
+    print x
+    print '~~~~~~~~~~~~~~~~'
     return x
 
->>>>>>> Stashed changes
 def solve_constrainedLSQ(A,b,C,d):
     """
     Solves the direct, constraint least squares problem ||Ax-b||_2 subject to Cx=d using 
@@ -203,10 +221,74 @@ def house(vec):
     v = v.T
     return v, beta
 
+
+# QR factorization via the Modified Gram Schmidt Method!
+def qr(A, thin=None):
+    """
+    Returns the QR factorization via the Modified Gram Schmidt Method
+
+    :param numpy matrix A: Matrix input for QR factorization
+    :param boolean thin: Set thin to 1 to compute a thin QR factorization. The default is a regular QR factorization.
+    :return: Q, the orthogonal matrix
+    :rtype: numpy matrix
+    :return: R, the upper triangular matrix
+    :rtype: numpy matrix
+
+    **Sample declaration**
+    :: 
+        >> Q, R = qr(A)
+    """
+    A = np.matrix(A)
+    m , n = A.shape
+    #u = np.min([m, n])
+    #h = np.max([m, n])
+
+    if thin == 1:
+        Q = np.mat(np.eye(m,n), dtype='float64')
+        R = np.mat(np.zeros((n, n)), dtype='float64')
+        u = n
+    else:
+        Q = np.mat(np.eye(m,m), dtype='float64')
+        R = np.mat(np.zeros((m, n)), dtype='float64')
+        u = n 
+
+
+    # Now loop!
+    for k in range(0, u):
+        
+        # re-orthogonalization
+        if k != 0:
+            for i in range(0, k-1):
[…]
