### 2016-12-16 781eff4
qr -- working temp.
