### 2016-12-16 ebc17f7
working qr!
