+++ effective_quadratures/qr.py
@@ -2,6 +2,8 @@
 """Utilities with QR factorization"""
 import numpy as np
 from utils import error_function
+np.set_printoptions(precision=4)
+np.set_printoptions(suppress=True)
 #****************************************************************************
 # Functions to code:
 #    
@@ -35,17 +37,12 @@ def solveCLSQ(A,b,C,d):
     :rtype: ndarray
 
     """
-    A = np.mat(A)
-    C = np.mat(C)
-   
     # Size of matrices!
     m, n = A.shape
     p, q = b.shape
     k, l = C.shape
     s, t = d.shape
 
-    print k, s
-    print m, n
     
     # Check that the number of elements in b are equivalent to the number of rows in A
     if m != p:
@@ -53,19 +50,10 @@ def solveCLSQ(A,b,C,d):
     elif k != s:
         error_function('ERROR: mismatch in sizes of C and d') 
 
-    notused , R = qr(C.T, thin=1)#qr_Householder(C.T, 1) # Thin QR factorization on C'
-    Q , notused = qr(C.T, thin=0)#qr_Householder(C.T)
-
-    print '***Debugging from qr.py****'
-    print Q
-    print R
-    print '~~~~~~~~~~~~~~~'
-    # up to here is OK! - Pranay Dec 16th 2016
-
+    #notused , R = qr(C.T, thin='yes') #qr_Householder(C.T, 1) # Thin QR factorization on C'
+    Q , R = qr_Householder(C.T)
+    
     u = np.linalg.inv(R.T) * d
-    print '---- Check u ----'
-    print u
-    print '####################'
     A_hat = A * Q
     z, w = A_hat.shape
 
@@ -73,23 +61,10 @@ def solveCLSQ(A,b,C,d):
     Ahat_1 = A_hat[:, 0:len(u)]
     Ahat_2 = A_hat[:, len(u) : w]
     r = b - Ahat_1 * u
-    print '***Check Ahat_1 and Ahat_2 ****'
-    print Ahat_1
-    print Ahat_2
-    print '~'
-    print A_hat
-    print r
-    print '$$$$$$$$$$$$$'
-    
-    
-    # Solve the least squares problem
-    v = solveLSQ(Ahat_2, r)
+    v = np.linalg.lstsq(Ahat_2, r)
     x = Q * np.vstack([u, v]) 
-    print '** Final steps ***'
-    print v
-    print x
-    print '~~~~~~~~~~~~~~~~'
-    return x
+  
+    return 0
 
 def solve_constrainedLSQ(A,b,C,d):
     """
@@ -176,7 +151,7 @@ def solveLSQ(A, b):
 
     """ 
     # Direct methods!
-    Q, R = qr_Householder(A, 1)
+    Q, R = qr_Householder(A)
     x = np.linalg.inv(R) * Q.T * b
     x = np.array(x)
     return x
@@ -223,9 +198,9 @@ def house(vec):
 
 
 # QR factorization via the Modified Gram Schmidt Method!
-def qr(A, thin=None):
+def qr_MGS(A):
     """
-    Returns the QR factorization via the Modified Gram Schmidt Method
+    Returns the thin QR factorization via the Modified Gram Schmidt Method
 
     :param numpy matrix A: Matrix input for QR factorization
     :param string thin: Set thin to 'yes' to compute a thin QR factorization. The default is a regular QR factorization, i.e., the string is set to 'no'.
@@ -240,22 +215,16 @@ def qr(A, thin=None):
     """
     A = np.matrix(A)
     m , n = A.shape
[…]
