### 2016-12-17 4d0e885
working on bidiagonalization
