+++ effective_quadratures/qr.py
@@ -7,23 +7,66 @@ np.set_printoptions(suppress=True)
 #****************************************************************************
 # Functions to code:
 #    
-# 1. Block & Block Recursive QR
+# 1. Block & Block Recursive QR (see GVL and Demmel)
 # 2. Constrained & weighted least squares - direct elimination & null space
 # 3. Singular value decomposition & bidiagonalization
 # 4. Routines for total least squares problem
 # 5. Rank 1 updates for QR factorization
 # 6. The 'Practical QR' algorithm
 # 7. Randomized QR 
-#****************************************************************************
-# SVD via QR!
-def tsvd(A):
-    return 0
 
-# Using Demmel's algorithm for QR
-def qrblocks(A, n):
-    # n is the number of blocks!
-    return 0
-    
+
+# Bidiagonalization, symmetric QR, CS decomposition, SVD, gSVD!
+#****************************************************************************
+def cs(A, B):
+    """
+    Computes the CS decomposition!
+    """
+
+
+def gsvd(A,B):
+    """
+    Computes a generalized singular value decomposition
+    """
+
+def svd(A):
+    """
+    Computes the singular value decomposition of a matrix!
+    """
+
+def bidiag(A):
+    """
+    Computes a bidiagionalization of a matrix
+
+    ** Notes **
+    Uses the algorithm of Golub and Kahan (1965) and requires 4mn^2 - 4n^3/3 flops.
+    """
+    m, n = A.shape
+    G = np.mat(np.eye(m,n), dtype='float64') # For storing U and V Householder vectors
+    for j in range(0, n):
+        v, beta = house(A[j:m, j])
+        A[j:m, j:n] = (np.identity(m-j) - np.multiply(beta,  v * v.T )) * A[j:m, j:n]
+        G[j+1:m, j] = v[1:m-j+1]
+        #print 'Pass'
+
+        if j <= n - 2 : 
+            v, beta = house(A[j,j+1:n].T)
+            A[j:m, j+1:n] = A[j:m, j+1:n] * (np.identity(n-j-1) -  np.multiply(beta,  v * v.T ) )
+            G[j,j+2:n] = v[1 : n-j].T
+            
+    # Unpack U and V -- using backward accumulation!?
+    for j in range(n-1, -1, -1):
+        if j >= m - 1:
+            if not A[j+1:m, j]:
+                beta = 2.0
+                Q[j:m, j:m] = Q[j:m, j:m] - ( np.multiply( beta,  Q[j:m, j:m] ) )
+        else:
+            v = np.mat(np.vstack([1.0, A[j+1:m, j] ]), dtype='float64')
+            beta =  2.0/(1.0 + np.linalg.norm(A[j+1:m, j], 2)**2)
+            Q[j:m, j:m] = Q[j:m, j:m] - ( ( np.multiply( beta,  v) ) * (v.T  * Q[j:m, j:m] ) )
+
+    return U, A, V
+
 def solveCLSQ(A,b,C,d):
     """
     Solves the direct, constraint least squares problem ||Ax-b||_2 subject to Cx=d using 
@@ -42,99 +85,23 @@ def solveCLSQ(A,b,C,d):
     p, q = b.shape
     k, l = C.shape
     s, t = d.shape
-
     
     # Check that the number of elements in b are equivalent to the number of rows in A
     if m != p:
-        error_function('ERROR: mismatch in sizes of A and b')
+        raise(ValueError, 'solveCLSQ(): mismatch in sizes of A and b')
     elif k != s:
-        error_function('ERROR: mismatch in sizes of C and d') 
[…]
