### 2016-12-18 680e98a
added bidiagionalization routine.
