### 2016-12-18 47aff16
added Givens and symmetric QR
