+++ effective_quadratures/qr.py
@@ -18,30 +18,142 @@ np.set_printoptions(suppress=True)
 
 # Bidiagonalization, symmetric QR, CS decomposition, SVD, gSVD!
 #****************************************************************************
-def cs(A, B):
-    """
-    Computes the CS decomposition!
-    """
-
-
 def gsvd(A,B):
     """
     Computes a generalized singular value decomposition
     """
+    m, n = A.shape
 
-def svd(A):
+
+    # Step 1. Compute a QR factorization of [A;B]
+    Q, R = qr_MGS(np.vstack([A, B]))
+
+    # Step 2. CS decomposition of 
+
+    return 0
+
+def svd(A, eps):
     """
-    Computes the singular value decomposition of a matrix!
+    Computes the singular value decomposition of a matrix based on Golub-Kahan
     """
+
+    # Step 1. Compute a bidiagionalization of the matrix A:
+    U, A, V = bidiag(A)
+
+    return 0
+
+
+def svd_step(B):
+    """
+    Computes a single step of the Golub-Kahan SVD iteration. 
+    """
+    # trailing eigenvalue
+    mu = T[n,n] - (T[n, n-1]**2)/(d + np.sign(d) * np.sqrt(d**2 + T[n,n-1]**2))
+    return 0
+
+    
+def cs(Q1, Q2):
+    """
+    Computes a cosine-sine decomposition
+    """
+    m, p = Q1.shape
+    n, pb = Q2.shape
+
+    if p != pb:
+        raise(ValueError, 'cs(): Number of columns in Q1 must be equivalent to number of columns in Q2')
+    
+    if m < n:
+        V, U, Z, S, C = cs(Q2, Q1)
+        j = range(p, 0, -1)
+        C = C[:, j]
+        S = S[:, j]
+        Z = Z[:, j]
+        m = np.min([m,p])
+        n = np.min([n, p])
+        i = range(m,0, -1)
+        C[0:m, :] = C[i,:]
+        U[:, 0:m] = U[:, i]
+        i = range(n,0, -1)
+        S[0:n, :] = S[i, :]
+        V[:, 0:n] = V[:, i]
+        return U, V, Z, C, S
+    
+    # Compute the svd
+    U, C, Z = np.linalg.svd(Q1)
+    q = np.min([m,p])
+    i = range(0, q, 1)
+    j = range(q, 0, -1)
+    C[i,i] = C[j,j]
+    U[:,i] = U[:,j]
+    Z[:,i] = Z[:,j]
+    S = Q2 * Z
+
+    if q == 1:
+        k = 0
+    elif m < p:
+        k = n
+    else:
+        entries = np.diag(C)
+        k = np.max([np.nonzero(entries <= 2)])
+    
+    V, R = qr_Householder(S[:, 0:k])
+    S = V.T * S
+    r = np.min([k,m])
+    S[:, 0:r] = np.diag(S[:, 0:r])
+    if m == 1 and p > 1:
+        S[0,0] = 0
+
+    if k < np.min([n,p])
+        r = np.min([n,p])
+        i = range(k+1, n, 1)
+        j = range(k+1, r, 1)
+
+        # Compute svd!
+        [UT, ST, VT] = np.linalg.svd(S[i, j])
+        
+        if k > 0:
+            S[1:k, j] = ST
+        
+        C[:, j] = C[:, j] * VT
+        V[:, i] = V[:, i] * UT
+        Z[:, j] = Z[:, j] * VT
+        i = range(k,q,1)
+        Q, R = qr(C[i,j])
+        C[i,j] = np.triu(np.tril(R))
[…]
