+++ README.md
@@ -12,17 +12,13 @@ www.effective-quadratures.org
 # Installation
 For installation on Mac and Linux systems, simply type the following into the terminal. 
 ```bash
-> sudo pip install effective-quadratures
-```
-or alternatively try
-```bash
 > sudo python setup.py install
 ```
 For installation on Windows, you will need [Anaconda] (https://www.continuum.io/downloads#windows); select the Python 2.7 option. Upon successful installation, open the Sypder IDE and go into the Effective-Quadratures-master directory and type the following in the command window
 ```bash
 > python setup.py install
 ```
-This should build the code. Just make sure you include the location of effective_quadratures folder to your python file and you should be good to go. 
+This should build the code. Just make sure you include the location of effective_quadratures folder to your python file and you should be good to go.
 
 # Testing
 Effective-Quadratures uses [Travis](https://travis-ci.org/Effective-Quadratures/Effective-Quadratures) for testing. Our objective is to ensure that regardless of the platform used, the modules will provide the same numerical answer. For testing the suite of tools type
@@ -39,4 +35,4 @@ Pranay Seshadri <br>
 *University of Cambridge* <br>
 
 # Funding
-This tool has been supported by funding from the Air Force Office of Scientific Research (AFOSR) under grant number FA9550-15-1-0018 and from the Engineering Physical Sciences Research Council (EPSRC) U.K. 
+This tool has been supported by funding from the Air Force Office of Scientific Research (AFOSR) under grant number FA9550-15-1-0018 and from the Engineering Physical Sciences Research Council (EPSRC) U.K.
