+++ .coveralls.yml
@@ -1,0 +1,2 @@
+service_name: travis-pro
+repo_token: eHcbY08k1mIpixsqTTQ1v9D5HQjrRVIAH
