### 2016-12-19 b7822e3
major clean-up
