+++ tests/test_indexset.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 from unittest import TestCase
 import unittest
 from effective_quadratures.indexset import IndexSet
