+++ ipython-notebooks/Vegetation_results/ke_output.txt
@@ -1,16 +1,0 @@
-Case_no, Mean KE (mts2/sec2), Mean_Bottom_KE(mt2/sec2): 
-1	, 436.770095328	, 3.36795747394	 
-2	, 524.746650218	, 5.44650880811	 
-3	, 411.681338502	, 2.91845873939	 
-4	, 445.17901124	, 3.32148807464	 
-5	, 450.962042491	, 3.82970156422	 
-6	, 498.159288373	, 4.53919746457	 
-7	, 449.866057827	, 3.35276281333	 
-8	, 392.910603292	, 2.80429730865	 
-9	, 430.181869637	, 3.11404648261	 
-10	, 416.961026324	, 2.99105005099	 
-11	, 500.449092724	, 4.54659673351	 
-12	, 475.765464167	, 4.21066789638	 
-13	, 509.803436682	, 4.84261138781	 
-14	, 456.077067392	, 3.77037741686	 
-15	, 494.391117514	, 4.48910980379	 
