+++ ipython-notebooks/Vegetation_results/grad_water_level_output.txt
@@ -1,16 +1,0 @@
-Case_no, Gradient of water level(mts.), istore, jstore: 
-1	, 0.000547886	 50	 55	 
-2	, 0.000407457	 45	 55	 
-3	, 0.000470981	 50	 55	 
-4	, 0.000536412	 49	 55	 
-5	, 0.000542	 49	 55	 
-6	, 0.000408232	 45	 55	 
-7	, 0.000477314	 50	 55	 
-8	, 0.000620097	 50	 55	 
-9	, 0.000508308	 50	 55	 
-10	, 0.000545859	 50	 55	 
-11	, 0.000407234	 45	 55	 
-12	, 0.000523135	 45	 55	 
-13	, 0.00040561	 45	 55	 
-14	, 0.000550494	 49	 55	 
-15	, 0.000473112	 45	 55	 
