+++ ipython-notebooks/Vegetation_results/dissip_output.txt
@@ -1,16 +1,0 @@
-Case_no, Mean wave dissipation: 
-1	, 0.00288641956093	 
-2	, 0.000588094010481	 
-3	, 0.00406629056334	 
-4	, 0.0030033701336	 
-5	, 0.00238950553151	 
-6	, 0.00147861895071	 
-7	, 0.00351944094428	 
-8	, 0.0035562196531	 
-9	, 0.00332515724102	 
-10	, 0.00382217771319	 
-11	, 0.00217241293763	 
-12	, 0.00125030030237	 
-13	, 0.00102605317042	 
-14	, 0.00160467600591	 
-15	, 0.0018514665637	 
