### 2016-12-21 39122ca
Regression notebook
