### 2016-12-21 a08e540
WIP on the quadratic regression fit.
