+++ effective_quadratures/qr.py
@@ -2,8 +2,8 @@
 """Utilities with QR factorization"""
 import numpy as np
 from utils import error_function
-#np.set_printoptions(precision=15)
-#np.set_printoptions(suppress=True)
+np.set_printoptions(precision=4)
+np.set_printoptions(suppress=True)
 #****************************************************************************
 # Functions to code:
 #    
@@ -18,46 +18,15 @@ from utils import error_function
 
 # Bidiagonalization, symmetric QR, CS decomposition, SVD, gSVD!
 #****************************************************************************
-def gsvd(A,B):
-    """
-    Computes a generalized singular value decomposition
-    """
-    m, n = A.shape
-
-
-    # Step 1. Compute a QR factorization of [A;B]
-    Q, R = qr_MGS(np.vstack([A, B]))
-
-    # Step 2. CS decomposition of 
-
-    return 0
-
-def svd(A, eps):
-    """
-    Computes the singular value decomposition of a matrix based on Golub-Kahan
-    """
-
-    # Step 1. Compute a bidiagionalization of the matrix A:
-    U, A, V = bidiag(A)
-
-    return 0
-
-
-def svd_step(B):
-    """
-    Computes a single step of the Golub-Kahan SVD iteration. 
-    """
-    # trailing eigenvalue
-    mu = T[n,n] - (T[n, n-1]**2)/(d + np.sign(d) * np.sqrt(d**2 + T[n,n-1]**2))
-    return 0
-
-    
 def cs(Q1, Q2):
     """
     Computes a cosine-sine decomposition
     """
     m, p = Q1.shape
     n, pb = Q2.shape
+    U = np.mat(np.zeros((m, m)), dtype='float64')
+    #V = np.mat(np.zeros((m,n))
+    # Allocate space for variables!
 
     if p != pb:
         raise(ValueError, 'cs(): Number of columns in Q1 must be equivalent to number of columns in Q2')
@@ -126,6 +95,29 @@ def cs(Q1, Q2):
 
        return 0 
 
+def gsvd(A,B):
+    """
+    Computes a generalized singular value decomposition
+    """
+
+    # Allocate memory
+    m, p = A.shape
+    n, p = B.shape
+    q = np.min([m+n, p])
+    U = np.mat(np.zeros((m, m)), dtype='float64')
+    V = np.mat(np.zeros((n, n)), dtype='float64')
+    X = np.mat(np.zeros((p, q)), dtype='float64')
+
+    # Step 1. Compute a QR factorization of [A;B]
+    Q, R = qr_MGS(np.vstack([A, B]))
+
+    # Step 2. CS decomposition of Q
+    U, V, Z, C, S = cs(Q[0:m, :], Q[m+1:m+n, :])
+
+    # Step 3. Compute X
+    X = R.T * Z
+
+    return U, V, X, C, S
 
 def implicitSymmetricQR(T):
     """
@@ -163,7 +155,6 @@ def givens(a, b):
     for scalars a and b.
 
     """
-    #G = np.mat(np.eye(2), dtype='float64')
     if b == 0:
         c = 1
         s = 1
@@ -176,14 +167,53 @@ def givens(a, b):
             tau = (-1.0 * b)/(a * 1.0)
             c = 1.0/np.sqrt(1 + tau**2)
             s = c * tau
-            
-    # Place values of c and s into G
-    #G[0,0] = c
-    #G[0,1] = s
-    #G[1,0] = -1.0 * s
-    #G[1,1] = c
 
     return c, s
+def qr_Givens(A):
+    """
+    Returns the Givens QR factorization of a matrix A 
+
+    :param numpy matrix A: Matrix input for QR factorization
+    :return: Q, the m x m orthogonal matrix
+    :rtype: numpy matrix
+    :return: R, the m x n upper triangular matrix
[…]
