+++ effective_quadratures/effectivequads.py
@@ -4,7 +4,7 @@ from parameter import Parameter
 from polynomial import Polynomial
 from qr import qr_MGS, solveLSQ, solveCLSQ
 from indexset import IndexSet
-from utils import error_function, evalfunction
+from utils import error_function, evalfunction, evalgradients
 import numpy as np
 
 class EffectiveSubsampling(object):
@@ -57,7 +57,53 @@ class EffectiveSubsampling(object):
             self.method = method
         else:
             self.method = 'QR'
-            
+
+        # Items to set!
+        A, quadrature_pts, quadrature_wts = getA(self)
+        self.A = A
+        self.tensor_quadrature_points = quadrature_pts
+        self.tensor_quadrature_weights = quadrature_wts
+        if uq_parameters[0].derivative_flag == 1:
+            self.C = getC(self)
+        else:
+            self.C = None
+        self.no_of_basis_terms = index_set.getCardinality() 
+        self.C_subsampled = None
+        self.A_subsampled = None
+        self.no_of_evals = None
+        self.subsampled_quadrature_points = None
+        self.subsampled_quadrature_weights = None # stored as a diagonal matrix??
+        self.row_indices = None
+        self.dimensions = len(uq_parameters)
+    
+    def set_no_of_evals(self, no_of_evals):
+
+        # Once the user provides the number of evaluations required, we can set a few items!
+        self.no_of_evals = no_of_evals
+        Asquare, esq_pts, W, row_indices = getSquareA(self)
+        self.A_subsampled = Asquare
+        self.subsampled_quadrature_points = esq_pts
+        self.subsampled_quadrature_weights = W
+        self.row_indices = row_indices
+
+        # If the user has turned on the gradient flag!
+        if self.C is not None:
+            dimensions = len(self.C)
+            C0 = self.C[0] # Which by default has to exist!
+            C0 = C0.T
+            rows, cols = C0.shape
+            C_subsampled = np.mat( np.zeros((dimensions*len(row_indices), cols)), dtype='float64')
+            counter = 0
+            for i in range(0, dimensions):
+                temp_matrix = self.C[i].T
+                for j in range(0, len(row_indices)):
+                    for k in range(0,cols):
+                        C_subsampled[counter,k] = temp_matrix[row_indices[j],k]
+                    counter = counter + 1 
+            self.C_subsampled = C_subsampled
+
+
+
     def getAmatrix(self):
         """
         Returns a matrix of multivariate orthogonal polynomials evaluated at a tensor grid of quadrature points.
@@ -92,18 +138,7 @@ class EffectiveSubsampling(object):
         Asquare, esq_pts, W, points = getSquareA(self, maximum_number_of_evals)
         return Asquare, esq_pts, W, points
     
-    def getCmatrix(self):
-        """
-        Returns the full C matrix
-        """
-        stackOfParameters = self.uq_parameters
-        polynomial_basis = self.index_set
-        dimensions = len(stackOfParameters)
-        polyObject_for_basis = Polynomial(stackOfParameters, polynomial_basis) 
-        points, weights = polyObject_for_basis.getPointsAndWeights()
-        not_used, C = polyObject_for_basis.getMultivariatePolynomial(points)
-        Cfull = cell2matrix(C)
-        return Cfull
+    
 
     # Method below will change shortly!    
     def getCsubsampled(self, maximum_number_of_evals):
@@ -147,7 +182,8 @@ class EffectiveSubsampling(object):
         Asquare, esq_pts, W, points = getSquareA(self, maximum_number_of_evals)
         return esq_pts
 
-    def solveLeastSquares(self, maximum_number_of_evals, function_values):
+
[…]
