+++ effective_quadratures/effectivequads.py
@@ -103,86 +103,6 @@ class EffectiveSubsampling(object):
             self.C_subsampled = C_subsampled
 
 
-
-    def getAmatrix(self):
-        """
-        Returns a matrix of multivariate orthogonal polynomials evaluated at a tensor grid of quadrature points.
-
-        :param EffectiveSubsampling object: An instance of the EffectiveSubsampling class
-        :return: A, an m-by-k matrix where m is the cardinality of the index set used to define the EffectiveSubsampling object
-            and k are the number of tensor grid quadrature points formed by the order prescribed when defining each Parameter in 
-            the EffectiveSubsampling object
-        :rtype: numpy matrix
-
-        **Sample declaration**
-        :: 
-            >> eq.getAMatrix()
-        """
-        return getA(self)
-
-    def getAsubsampled(self, maximum_number_of_evals):
-        """
-        Returns a matrix of multivariate orthogonal polynomials evaluated at a subsample of the tensor grid of quadrature points.
-
-        :param EffectiveSubsampling object: An instance of the EffectiveSubsampling class
-        :param integer maximum_number_of_evals: The maximum number of evaluations the user would like. This value has to be atleast equivalent to the
-            total number of basis terms of the index set.
-        :return: A, an m-by-k matrix where m is the cardinality of the index set used to define the EffectiveSubsampling object
-            and k are the number of subsamples given by the integer maximum_number_of_evals.
-        :rtype: numpy matrix
-        
-        **Sample declaration**
-        :: 
-            >> eq.getASubsampled()
-        """
-        Asquare, esq_pts, W, points = getSquareA(self, maximum_number_of_evals)
-        return Asquare, esq_pts, W, points
-    
-    
-
-    # Method below will change shortly!    
-    def getCsubsampled(self, maximum_number_of_evals):
-        """
-        Returns a matrix of multivariate derivative orthogonal polynomials evaluated at a set of quadrature points.
-
-        :param EffectiveSubsampling object: An instance of the EffectiveSubsampling class
-        :param ndarray quadrature_subsamples: The quadrature points are which the matrix of derivative orthogonal polynomials must be evaluated at.
-        :return: C, a cell with d matrices of size m-by-l matrix where m is the cardinality of the index set used to define the EffectiveSubsampling object
-            and l is the number of points in quadrature_subsamples. The integer d represents the dimensionality of the problem, equivalent to the number of
-            parameters defined when initializing the constructor.
-        :rtype: numpy matrix
-
-        **Sample declaration**
-        :: 
-            >> eq.getCSubsampled()
-        """
-        stackOfParameters = self.uq_parameters
-        polynomial_basis = self.index_set
-        dimensions = len(stackOfParameters)
-        polyObject_for_basis = Polynomial(stackOfParameters, polynomial_basis) 
-        Asquare, esq_pts, W, points = getSquareA(self, maximum_number_of_evals)
-        not_used, C = polyObject_for_basis.getMultivariatePolynomial(esq_pts)
-        #Cfull = cell2matrix(C)
-        return C
-
-    def getEffectivelySubsampledPoints(self, maximum_number_of_evals, flag=None):
-        """
-        Returns the effectively subsampled quadrature points. See this `paper <https://arxiv.org/abs/1601.05470>`_, for further details. 
-
-        :param EffectiveSubsampling object: An instance of the EffectiveSubsampling class
-        :param integer maximum_number_of_evals: The maximum number of evaluations the user would like. This value has to be atleast equivalent to the
-            total number of basis terms of the index set.        
-        :return: esq_pts, a maximum_number_of_evals-by-d matrix of quadrature points, where d represents the dimensionality of the problem.
-        :rtype: numpy matrix
-
-        **Sample declaration**
[…]
