+++ effective_quadratures/qr.py
@@ -206,25 +206,31 @@ def solveCLSQ(A,b,C,d):
     #    raise(ValueError, 'solveCLSQ(): mismatch in sizes of C and d') 
 
     # Uses the procedure prescribed by Bjorck
-    Q, R, pvec = qr_MGS(C, pivoting=True)
-    m1, n1 = R.shape
-    P = permvec2mat(pvec)
-    r = np.linalg.matrix_rank(C)
+    if method is 'weighted':
+        
+    if method is 'equality':
+        Q, R, pvec = qr_MGS(C, pivoting=True)
+        m1, n1 = R.shape
+        P = permvec2mat(pvec)
+        r = np.linalg.matrix_rank(C)
+        R_11 = R[0:r, 0:r]
+        R_12 = R[0:r, r:n1]
+        d_tilde = Q.T * d
+        d1_tilde = d_tilde[0 : r]
+        d2_tilde = d_tilde[r: m1]
+        A_tilde = A * P
+        A1_tilde = A_tilde[:, 0 : r]
+        A2_tilde = A_tilde[:, r : m1]
+        A2_hat = A2_tilde - A1_tilde * np.linalg.inv(R_11) * R_12
+        b_hat = b - A1_tilde * np.linalg.inv(R_11) * d1_tilde
+        x2_tilde = solveLSQ(A2_hat, b_hat)
+        x1_tilde = np.linalg.inv(R_11) * (d1_tilde - R_12 * x2_tilde)
+        x_tilde = np.mat( np.vstack([x1_tilde, x2_tilde]) , dtype='float64')
+        x = P * x_tilde
 
-    R_11 = R[0:r, 0:r]
-    R_12 = R[0:r, r:n1]
-    d_tilde = Q.T * d
-    d1_tilde = d_tilde[0 : r]
-    d2_tilde = d_tilde[r: m1]
-    A_tilde = A * P
-    A1_tilde = A_tilde[:, 0 : r]
-    A2_tilde = A_tilde[:, r : m1]
-    A2_hat = A2_tilde - A1_tilde * np.linalg.inv(R_11) * R_12
-    b_hat = b - A1_tilde * np.linalg.inv(R_11) * d1_tilde
-    x2_tilde = solveLSQ(A2_hat, b_hat)
-    x1_tilde = np.linalg.inv(R_11) * (d1_tilde - R_12 * x2_tilde)
-    x_tilde = np.mat( np.vstack([x1_tilde, x2_tilde]) , dtype='float64')
-    x = P * x_tilde
+    if method is 'inequality':
+        # call CVXOPT!???
+        
     return x
 
 def solveLSQ(A, b):
@@ -504,4 +510,4 @@ def main():
 
     x = solveCLSQ(A, b, C, d)
 
-main()
+#main()
