### 2017-01-02 35518cb
WIP with qr
