### 2017-01-04 ec8a570
Delete btest_equadII.py
