### 2017-01-04 #126 @katyhuff
Sorry for the delay. I have tried to contact a few reviewers offline, but due to the holidays here, there's been some delay. So, I'll try to contact some reviewers here in the github issue.

Dear @moorepants , @jordigh , @nicoguaro :

I saw that you were on the JOSS reviewers list and this paper seemed to be something any of the three of you would be an excellent reviewer for. Would any of you be interested  in conducting the review of this JOSS paper? If you've not reviewed a JOSS paper yet, please note that it should be a short-ish process. This is how to start: http://joss.theoj.org/about#reviewer_guidelines .

What do you think? If I can assign you as a reviewer for this paper, please let me know.
