### 2017-01-05 2f49b48
added error estimator
