+++ tests/test_equadII.py
@@ -4,7 +4,7 @@ from effective_quadratures.polynomial import Polynomial
 from effective_quadratures.indexset import IndexSet
 from effective_quadratures.computestats import Statistics
 from effective_quadratures.effectivequads import EffectiveSubsampling
-from effective_quadratures.utils import meshgrid, twoDgrid, evalfunction, evalgradients
+from effective_quadratures.utils import compute_errors, evalfunction, evalgradients
 import numpy as np
 import matplotlib.pyplot as plt
 import math
@@ -17,29 +17,29 @@ def fun(x):
 def fungrad(x):
     return [ 4*x[0] - 5*x[1] , -5*x[0] + 9*x[1]**2 ] 
 
-#def fungradnoised(x):
-#    return [np.exp(x[0] + x[1]) - 0.05*np.random.randn(1), np.exp(x[0] + x[1])- 0.05*np.random.randn(1)] 
 os.system('clear')
+
+###############################################################################################
+# Without gradients and with as many basis function as possible!
+################################################################################################
+value = 60
+x1 = Parameter(param_type="Uniform", lower=-1, upper=1, points=value)
+x2 = Parameter(param_type="Uniform", lower=-1, upper=1, points=value)
+uq = Polynomial([x1,x2])
+all_coefficients, all_indices, evaled_pts = uq.getPolynomialCoefficients(fun)
+print len(all_coefficients)
+
+################################################################################
+# Now with gradients and reducing the number of basis terms
+################################################################################
 value = 12
 x1 = Parameter(param_type="Uniform", lower=-1, upper=1, points=value, derivative_flag=1)
 x2 = Parameter(param_type="Uniform", lower=-1, upper=1, points=value, derivative_flag=1)
 parameters = [x1, x2]
 hyperbolic_cross = IndexSet("Hyperbolic basis", orders=[value-1,value-1], q=1.0)
-
-print hyperbolic_cross.getIndexSet()
-
-
-###################
-# Without gradients!
-###################
 esq = EffectiveSubsampling(parameters, hyperbolic_cross)
-esq.set_no_of_evals(esq.no_of_basis_terms) # in effect making it a square system!
-x = esq.computeCoefficients(fun)
 print 'Number of basis terms desired in expansion: '+'\t'+str(esq.no_of_basis_terms)
 
-#######################
-# Now with gradients
-#######################
 # 1. Determine least number of subsamples required!
 minimum_subsamples = esq.least_no_of_subsamples_reqd()
 print 'Number of evaluations used (not counting gradients): '+'\t'+str(minimum_subsamples)
@@ -64,10 +64,11 @@ print '----------------------------'
 # 5. Compute coefficients using the two techniques!
 x1 =  esq.computeCoefficients(fun_values, grad_values, 'weighted')
 x1b =  esq.computeCoefficients(fun_values, grad_values, 'equality')
-#x1c =  esq.computeCoefficients(fun_values, grad_values, 'equality2')
+weighted_error = compute_errors(all_coefficients, all_indices, x1, esq.index_set.getIndexSet() )
+equality_error = compute_errors(all_coefficients, all_indices, x1b, esq.index_set.getIndexSet() )
 
-weighted_error = np.linalg.norm(x-x1, 2)
-equality_error = np.linalg.norm(x-x1b, 2)
+weighted_error = np.linalg.norm(weighted_error, 2)
+equality_error = np.linalg.norm(equality_error, 2)
 #equality_error2 = np.linalg.norm(x-x1c, 2)
 
 # 6. Compute 2-norm errors!
@@ -75,10 +76,8 @@ print 'Errors:'
[…]
