+++ effective_quadratures/effectivequads.py
@@ -102,17 +102,22 @@ class EffectiveSubsampling(object):
                     counter = counter + 1 
             self.C_subsampled = C_subsampled
 
-    def prune(self, number_of_columns_to_delete):
-        
+    def prune(self, number_of_columns_to_delete):  
         A = self.A_subsampled
         m, n = A.shape
-        C = self.C_subsampled
-        p, q = C.shape
         A_pruned = A[0:m, 0 : (n - number_of_columns_to_delete)]
-        C_pruned = C[0:p, 0 : (q - number_of_columns_to_delete)]
         self.A_subsampled = A_pruned
-        self.C_subsampled = C_pruned
+        self.index_set
+
+        # If clause for gradient case!
+        if self.C is not None:
+            C = self.C_subsampled
+            p, q = C.shape
+            C_pruned = C[0:p, 0 : (q - number_of_columns_to_delete)]
+            self.C_subsampled = C_pruned
+
         self.no_of_basis_terms = self.no_of_basis_terms - number_of_columns_to_delete
+        self.index_set.prune(number_of_columns_to_delete)
     
     def least_no_of_subsamples_reqd(self):
         k = 1
