### 2017-01-12 a050c5a
updated docs
