+++ documentation/source/conf.py
@@ -1,4 +1,4 @@
-import sphinx_rtd_theme
+import sphinx_bootstrap_theme
 import sys
 import os
 import shlex
@@ -36,14 +36,14 @@ import shlex
 sys.path.insert(0, os.path.abspath('../..'))
 html_theme = "sphinx_rtd_theme"
 source_suffix = '.rst'
-html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
+#html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 extensions = [
     'sphinx.ext.autodoc',
 ]
 master_doc = 'index'
 
 # General information about the project.
-project = u'Effective Quadratures'
+project = u' '
 copyright = u'2016, Pranay Seshadri'
 author = u'Pranay Seshadri'
 
@@ -56,224 +56,79 @@ version = '1.0'
 # The full version, including alpha/beta/rc tags.
 release = version
 
-# The language for content autogenerated by Sphinx. Refer to documentation
-# for a list of supported languages.
-#
-# This is also used if you do content translation via gettext catalogs.
-# Usually you set "language" from the command line for these cases.
 language = None
 
-# There are two options for replacing |today|: either, you set today to some
-# non-false value, then it is used:
-#today = ''
-# Else, today_fmt is used as the format for a strftime call.
-#today_fmt = '%B %d, %Y'
 
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-exclude_patterns = []
-pygments_style = 'sphinx'
-# The reST default role (used for this markup: `text`) to use for all
-# documents.
-#default_role = None
+html_theme = 'bootstrap'
+html_theme_path = sphinx_bootstrap_theme.get_html_theme_path()
+html_logo = "logo.png"
+html_theme_options = {
+    # Navigation bar title. (Default: ``project`` value)
+    'navbar_title': "",
 
-# If true, '()' will be appended to :func: etc. cross-reference text.
-#add_function_parentheses = True
+    # Tab name for entire site. (Default: "Site")
+    'navbar_site_name': "Docs",
 
-# If true, the current module name will be prepended to all description
-# unit titles (such as .. function::).
-#add_module_names = True
+    # A list of tuples containing pages or urls to link to.
+    # Valid tuples should be in the following forms:
+    #    (name, page)                 # a link to a page
+    #    (name, "/aa/bb", 1)          # a link to an arbitrary relative url
+    #    (name, "http://example.com", True) # arbitrary absolute url
+    # Note the "1" or "True" value above as the third argument to indicate
+    # an arbitrary url.
 
-# If true, sectionauthor and moduleauthor directives will be shown in the
-# output. They are ignored by default.
-#show_authors = False
+    # Render the next and previous page links in navbar. (Default: true)
+    'navbar_sidebarrel': True,
 
+    # Render the current pages TOC in the navbar. (Default: true)
+    'navbar_pagenav': True,
 
-# A list of ignored prefixes for module index sorting.
-#modindex_common_prefix = []
+    # Tab name for the current pages TOC. (Default: "Page")
+    'navbar_pagenav_name': "Page",
 
-# If true, keep warnings as "system message" paragraphs in the built documents.
-#keep_warnings = False
+    # Global TOC depth for "site" navbar tab. (Default: 1)
+    # Switching to -1 shows all levels.
+    'globaltoc_depth': 2,
 
-# If true, `todo` and `todoList` produce output, else they produce nothing.
-todo_include_todos = False
+    # Include hidden TOCs in Site navbar?
+    #
+    # Note: If this is "false", you cannot have mixed ``:hidden:`` and
+    # non-hidden ``toctree`` directives in the same page, or else the build
+    # will break.
+    #
[…]
