### 2017-01-18 44eaba6
added a notebook on where effective quadratures can fail.
