### 2017-01-18 #166 @whedon
[REVIEW]: Effective-Quadratures: Polynomials for Computational Engineering Studies
**Submitting author:** @psesh (<a href="http://orcid.org/0000-0002-7351-012X">Pranay Seshadri</a>)
**Repository:** <a href="https://github.com/Effective-Quadratures/Effective-Quadratures" target ="_blank">https://github.com/Effective-Quadratures/Effective-Quadratures</a>
**Version:** v5.2
