### 2017-01-18 #126 @arfon
@nicoguaro - please carry out your review over in #166 where the review checklist is available.
