+++ effective_quadratures/effectivequads.py
@@ -9,22 +9,31 @@ import numpy as np
 
 class EffectiveSubsampling(object):
     """
-    This class defines an EffectiveSubsampling object. Below are details of its constructor.
-
-    :param Parameter (an array of): A list of parameter objects.
+    :param array uq_parameters: A list of Parameter objects.
     :param IndexSet index_set: Polynomial index set. If an index set is not given, the constructor uses a tensor grid basis of polynomials. For total order and hyperbolic index sets, the user needs to explicity input an index set.
     :param string method: Subsampling strategy; options include: 'QR', which is the default option and 'Random'. See this `paper <https://arxiv.org/abs/1601.05470>`_, for further details. 
         We will be adding a 'Hybrid' strategy shortly that combines both QR and randomized techniques.
     
     Attributes:
-        * *self.A* (numpy matrix) : Matrix of the multivariate orthogonal polynomial (defined as per the index set basis) evaluated at all points of a tensor grid.
-        * *self.C* (cell) : A cell of d-dimensional numpy matrices of the derivative of the orthogonal polynomial evaluated at all points of a tensor grid.
-        
-    **Notes** 
-    
-    For the exact definitions of A, C, A_subsampled and C_subsampled please see: `paper <https://arxiv.org/abs/1601.05470>`_. The variables shown above are attributes of the constructor.
+        * **self.A**: (numpy matrix) Matrix of the multivariate orthogonal polynomial (defined as per the index set basis) evaluated at all points of a tensor grid.
+        * **self.C**: (cell) A cell of d-dimensional numpy matrices of the derivative of the orthogonal polynomial evaluated at all points of a tensor grid.
+        * **self.index_set**:(numpy array) An array of the indices used in the multivariate Polynomial. This changes when the user prunes down the number of columns.
+        * **self.tensor_quadrature_points**: (numpy matrix) Full tensor grid quadrature points
+        * **self.tensor_quadrature_weights**: (numpy matrix) Full tensor grid quadrature weights
+        * **self.A_subsampled**: (numpy matrix) Subsampled rows of A, obtained either via QR column pivoting or randomization
+        * **self.C_subsampled**: (numpy matrix) A stacked matrix comprising of subsampled rows of each matrix in the self.C
+        * **self.no_of_basis_terms**: (int) The number of columns in A_subsampled
+        * **self.dimension**: (int) The number of dimensions of the Polynomial
+        * **self.no_of_evals**: (int) The minimum number of model evaluations required to estimate the coefficients of the Polynomial
+        * **self.row_indices**: (int) The rows of A that are used in construction A_subsampled
+        * **self.subsampled_quadrature_points**: (numpy matrix) Subsampled quadrature points
+        * **self.subsampled_quadrature_weights**: (numpy matrix) Subsampled quadrature weights
 
-    **Sample declarations** 
+    **Notes:** 
+    
+    For the exact definitions of A, C, A_subsampled and C_subsampled please see: `paper <https://arxiv.org/abs/1601.05470>`_. 
+
+    **Sample usage:** 
     ::
         
         >> var1 = Parameter(points=12, shape_parameter_A=0.5, param_type='Exponential')
@@ -33,6 +42,8 @@ class EffectiveSubsampling(object):
         >> print eq.A
         >> print eq.dimensions
         >> print eq.subsampled_quadrature_points
+        >> print eq.no_of_evals
+        >> print esq.C_subsampled
     """
 
     # Constructor
@@ -57,7 +68,7 @@ class EffectiveSubsampling(object):
                 if orders_to_use[u] <= index_set.orders[u] :
                     count = count + 1
             if count > 0:
-                error_function('IndexSet: Basis orders: Ensure that the basis order is always -1 the number of points!')
[…]
