+++ WIP/WIP_test_spherical.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 import Ylm
 import numpy as np
 import matplotlib.pyplot as plt
@@ -11,6 +12,7 @@ def Pranay1():
     fig = plt.figure()
     ax = fig.add_subplot(111, projection='3d')
     ax.scatter(x,y,z,s=20*w)
+    plt.show()
 
 
 #Demo 2:
@@ -21,7 +23,8 @@ def Pranay2():
     fig = plt.figure()
     ax = fig.add_subplot(111, projection='3d')
     ax.scatter(x,y,z,s=20*w)
-        
+    plt.show()
+    
 #Demo 3: check SH integration accuracy
 def Pranay3():
     w,theta,phi = Ylm.gen_tri_quad(16)
@@ -29,4 +32,6 @@ def Pranay3():
     err1, err2 = Ylm.test_quads(x,y,z,w,5)
     plt.plot(err1)
     plt.plot(err2)
+
     
+Pranay2()
