+++ effective_quadratures/plots.py
@@ -34,3 +34,7 @@ def pcolor_plot():
     plt.colorbar()
     plt.xlim(0,max_order)
     plt.ylim(0,max_order)
+
+
+def plot_indexset(indexset):
+    indices = indexset.getIndexSet()
