### 2017-01-29 b9f0257
fixed sorting index sets
