### 2017-01-29 56e48a6
working isotropic Euclidean degree
