+++ effective_quadratures/qr.py
@@ -365,7 +365,6 @@ def qr_MGS(A, pivoting=None):
             
     return Q, R
 
-
 def qr_MGS_Pivoting(A):
 
     # Determine the size of
@@ -509,42 +508,3 @@ def rowNormalize(A):
         Normalization[i,i] = row_norms[i]
     A_normalized = np.dot(Normalization, A)
     return A_normalized, Normalization
-    
-def main():
-    A = np.mat(  [ [2.41681886, -3.60476229,  3.30915575, -1.94864415, -3.60476229,  5.37661775,  -4.93571118,  3.30915575, -4.93571118, -1.94864415],
-    [2.41681886,  3.60476229,  3.30915575,  1.94864415, -3.60476229, -5.37661775,  -4.93571118,  3.30915575,  4.93571118, -1.94864415],
-    [2.41681886, -3.60476229,  3.30915575, -1.94864415,  3.60476229, -5.37661775,  4.93571118,  3.30915575, -4.93571118,  1.94864415],
-    [2.41681886,  3.60476229,  3.30915575,  1.94864415,  3.60476229,  5.37661775,  4.93571118,  3.30915575,  4.93571118,  1.94864415],
-    [5.45328194,  3.21124382, -3.98276649, -5.94042304,  3.21124382,  1.8909873, -2.34530956, -3.98276649, -2.34530956, -5.94042304]], dtype='float64' )
-
-    C = np.mat( [ [ 0.0  ,        -0.0     ,      0.0 ,         -0.0      ,     1.73205081,  -2.58340893 ,  2.37155792,  -5.77667799 ,  8.61609918 , 10.74616371],
-    [0.0         ,  0.0 ,          0.0     ,      0.0 ,          1.73205081 , 2.58340893 ,  2.37155792 , -5.77667799 , -8.61609918  ,10.74616371],
-    [0.0        ,  -0.0 ,          0.0     ,     -0.0 ,          1.73205081 ,-2.58340893 ,  2.37155792 ,  5.77667799 , -8.61609918  ,10.74616371],
-    [0.0       ,    0.0 ,          0.0     ,      0.0 ,          1.73205081 , 2.58340893 ,  2.37155792 ,  5.77667799 ,  8.61609918  ,10.74616371],
-    [0.0      ,     0.0 ,         -0.0     ,     -0.0 ,          1.73205081 , 1.01994313 , -1.26499125 ,  2.28066217 ,  1.34300086  ,-1.67501636],
-    [0.0     ,      1.73205081,  -5.77667799,  10.74616371,  -0.0  ,        -2.58340893,  8.61609918 ,  0.0  ,         2.37155792  ,-0.0        ],
-    [0.0    ,       1.73205081,   5.77667799,  10.74616371,  -0.0  ,        -2.58340893, -8.61609918 ,  0.0  ,         2.37155792  ,-0.0        ],
-    [0.0   ,        1.73205081,  -5.77667799,  10.74616371,   0.0 ,          2.58340893,  -8.61609918,   0.0 ,          2.37155792 ,  0.0        ],
-    [0.0  ,         1.73205081,   5.77667799,  10.74616371,   0.0 ,          2.58340893,   8.61609918,   0.0 ,          2.37155792 ,  0.0        ],
-    [0.0 ,          1.73205081,   2.28066217,  -1.67501636,   0.0 ,          1.01994313,  1.34300086,  -0.00 ,        -1.26499125 , -0.0        ]], dtype='float64')
-
-    b = np.mat ([ [ -0.81440387],
-    [3.96388573],
-    [-0.81440387],
-    [3.96388573],
-    [8.56996634]], dtype='float64')
-
[…]
