### 2017-01-29 48c5b5d
issue with numpy version
