### 2017-01-29 b71577e
stats class in eq
