### 2017-01-30 1ff03f7
new plotting tools
