### 2017-01-31 bbd42f1
plotting utility
