+++ effective_quadratures/plotting.py
@@ -1,0 +1,195 @@
+#!/usr/bin/env python
+import matplotlib.pyplot as plt
+from matplotlib import rc
+import numpy as np
+import matplotlib as mpl
+
+
+def bestfit(x_train, y_train, x_test, y_test, x_label, y_label, filename=None):
+    opacity = 0.8
+    plt.rc('text', usetex=True)
+    plt.rc('font', family='serif')
+    mpl.rcParams['axes.linewidth'] = 2.0
+    fig = plt.figure()
+    ax = fig.add_subplot(1,1,1)
+    plt.grid()
+    ax.set_axis_bgcolor('whitesmoke')
+    plt.scatter(x_train, y_train, marker='o', s=120, alpha=opacity, color='orangered',linewidth=1.5)
+    plt.plot(x_test, y_test, linestyle='-', linewidth=2, color='steelblue')
+    ax.set_axisbelow(True)
+    adjust_spines(ax, ['left', 'bottom'])
+    plt.xlabel(x_label, fontsize=16)
+    plt.ylabel(y_label, fontsize=16)
+    plt.grid(b=True, which='major', color='w', linestyle='-', linewidth=2)
+    plt.grid(b=True, which='minor', color='w', linestyle='-', linewidth=2)
+    plt.xticks(fontsize=16)
+    plt.yticks(fontsize=16)
+    plt.tight_layout()
+    if filename is not None:
+        plt.savefig(filename, format='eps', dpi=300, bbox_inches='tight')
+    else:
+        plt.show()
+
+def lineplot(x, y, x_label, y_label, filename=None):
+    opacity = 0.8
+    plt.rc('text', usetex=True)
+    plt.rc('font', family='serif')
+    mpl.rcParams['axes.linewidth'] = 2.0
+    fig = plt.figure()
+    ax = fig.add_subplot(1,1,1)
+    plt.grid()
+    ax.set_axis_bgcolor('whitesmoke')
+    plt.plot(x, y, linestyle='-', linewidth=3, color='deepskyblue')
+    ax.set_axisbelow(True)
+    adjust_spines(ax, ['left', 'bottom'])
+    plt.xlabel(x_label, fontsize=16)
+    plt.ylabel(y_label, fontsize=16)
+    plt.grid(b=True, which='major', color='w', linestyle='-', linewidth=2)
+    plt.grid(b=True, which='minor', color='w', linestyle='-', linewidth=2)
+    plt.xticks(fontsize=16)
+    plt.yticks(fontsize=16)
+    plt.tight_layout()
+    if filename is not None:
+        plt.savefig(filename, format='eps', dpi=300, bbox_inches='tight')
+    else:
+        plt.show()
+
+def scatterplot(x, y, z, filename=None):
+    m, n = x.shape
+    p, q = y.shape
+    if n > m :
+        raise(ValueError, 'scatterplot(x, y): Matrix x of size m-by-n, must satisfy m>=n')
+    if m is not p:
+        raise(ValueError, 'scatterplot(x, y): The number of rows in x must be equivalent to the number of rows in y')
+    
+        opacity = 0.8
+        plt.rc('text', usetex=True)
+        plt.rc('font', family='serif')
+        mpl.rcParams['axes.linewidth'] = 2.0
+        fig = plt.figure()
+        ax = fig.add_subplot(1,1,1)
+        plt.grid()
+        ax.set_axis_bgcolor('whitesmoke')
+        plt.scatter(x, y, marker='s', s=70, alpha=opacity, color='limegreen',linewidth=1.5)
+        ax.set_axisbelow(True)
+        adjust_spines(ax, ['left', 'bottom'])
+        plt.xlabel(x_label, fontsize=16)
+        plt.ylabel(y_label, fontsize=16)
[…]
