### 2017-01-31 decd7b3
updates to plotting routines
