### 2017-01-31 b8edd0e
working Euclidean! (prior to sorting)
