+++ effective_quadratures/plotting.py
@@ -6,27 +6,6 @@ import matplotlib as mpl
 from mpl_toolkits.mplot3d import Axes3D
 from matplotlib import cm
 
-# A simple utility that helps to visualize plotting sparse and tensor grid coefficients
-def twoDgrid(coefficients, index_set):
-
-    # First determine the maximum tensor grid order!
-    max_order = int( np.max(index_set) ) + 1
-
-    # Now create a tensor grid with this max. order
-    y, x = np.mgrid[0:max_order, 0:max_order]
-    z = (x*0 + y*0) + float('NaN')
-
-    for i in range(0, max_order):
-        for j in range(0, max_order):
-            x_entry = x[i,j]
-            y_entry = y[i,j]
-            for k in range(0, len(index_set)):
-                if(x_entry == index_set[k,0] and y_entry == index_set[k,1]):
-                    z[i,j] = coefficients[k]
-                    break
-
-    return x,y,z, max_order
-
 def coeffplot2D(coefficients, index_set, x_label, y_label, filename=None):
     elements = index_set.elements
     x, y, z, max_order = twoDgrid(coefficients, elements)
@@ -40,7 +19,7 @@ def coeffplot2D(coefficients, index_set, x_label, y_label, filename=None):
     ax = fig.add_subplot(1,1,1)
     plt.grid()
     ax.set_axis_bgcolor('whitesmoke')
-    plt.pcolor(y,x, Zm, cmap='jet', vmin=-16, vmax=1)
+    plt.pcolor(y,x, Zm, cmap= cm.terrain, vmin=-16, vmax=1)
     plt.xlim(0, max_order)
     plt.ylim(0, max_order)
     ax.set_axisbelow(True)
@@ -86,44 +65,15 @@ def bestfit(x_train, y_train, x_test, y_test, x_label, y_label, filename=None):
     else:
         plt.show()
 
-def lineplot(x, y, x_label, y_label, filename=None):
-    opacity = 0.8
-    plt.rc('text', usetex=True)
-    plt.rc('font', family='serif')
-    mpl.rcParams['axes.linewidth'] = 2.0
-    fig = plt.figure()
-    ax = fig.add_subplot(1,1,1)
-    plt.grid()
-    ax.set_axis_bgcolor('whitesmoke')
-    plt.plot(x, y, linestyle='-', linewidth=3, color='deepskyblue')
-    ax.set_axisbelow(True)
-    adjust_spines(ax, ['left', 'bottom'])
-    plt.xlabel(x_label, fontsize=16)
-    plt.ylabel(y_label, fontsize=16)
-    plt.grid(b=True, which='major', color='w', linestyle='-', linewidth=2)
-    plt.grid(b=True, which='minor', color='w', linestyle='-', linewidth=2)
-    plt.xticks(fontsize=16)
-    plt.yticks(fontsize=16)
-    if filename is not None:
-        plt.savefig(filename, format='eps', dpi=300, bbox_inches='tight')
-    else:
-        plt.show()
-
-def scatterplot3D(x, y, z, x_label, y_label, z_label, filename=None):
-    x = np.mat(x)
-    y = np.mat(y)
-    m, n = x.shape
-    p, q = y.shape
-
-    print '***********'
-    print m, n
-    print p, q
-
+def bestfit3D(x_train, y_train, x_test, y_test, x_label, y_label, z_label, filename=None):
+    m, n = x_train.shape
+    p, q = y_train.shape
     if n > m :
-        raise(ValueError, 'scatterplot(x, y): Matrix x of size m-by-n, must satisfy m>=n')
[…]
