### 2017-02-01 eaabeb9
editing python notebooks.
