### 2017-02-05 #166 @nicoguaro
@arfon, the beginning of semester took my time. But I already tested the software.

>Effective-Quadratures (EQ) is a suite of tools for generating polynomials for parametric computational studies.

The inclusion of a statement of need is important, since it shows to a broader audience the usefulness of the software. Regarding the installation, it has dependencies on ``numpy`` and ``scipy``, although this is not stated in the README. It mentions that it should use Python 2.7, but there is not apparent reason for not supporting Python 3.

Some particular comments below.

- [ ] The version in the download section of the repository is 4 and not 3 as stated in submission.

- [ ] Functionality: Have the functional claims of the software been confirmed? Although there is a general description of what the software does, it would be more clear with the **statement of need**.
- [x] Performance: Have any performance claims of the software been confirmed? There are no claims about the performance.

- [ ] A statement of need. There is not a statement of need.
- [x] Example of use. This is done as a set of Jupyter Notebooks. I would suggest to include a simple example in the README.
- [ ] Functionality documentation. Not completely, see comment below.
- [ ] Community guidelines. There are not community guidelines stated, just contact information.

Regarding the documentation, I think that it should include more than just the docstrings from the code. Some points that I consider pertinent are below. Some examples (http://barbagroup.github.io/pygbe/docs/, http://cvxopt.org/documentation/index.html)
- Description of the software.
- Installation.
- How to use the software, examples. There is no need to repeat the notebooks, but maybe one or two examples. Maybe a toy example:  [Polynomial regression in 1D](http://nbviewer.jupyter.org/github/Effective-Quadratures/Effective-Quadratures/blob/master/ipython-notebooks/Regression1D.ipynb)..
- Link to the repository.
- FAQ.

- [ ] A statement of need. This part is missing.
- [ ] References. The paper does not have references.
