### 2017-02-05 #166 @psesh
@nicoguaro thank you for the review :)
@katyhuff I will address these comments and get back to you shortly.
