+++ effective_quadratures/parameter.py
@@ -4,12 +4,11 @@ import numpy as np
 from scipy.special import gamma
 import analyticaldistributions as analytical
 from utils import error_function
-from plotting import histogram, lineplot
+import matplotlib.pyplot as plt
 class Parameter(object):
     
     """
     This class defines a univariate parameter. Below are details of its constructor.
-
     :param double lower: Lower bound for the parameter. 
     :param double upper: Upper bound for the parameter
     :param integer points: Number of quadrature points to be used for subsequent
@@ -30,18 +29,14 @@ class Parameter(object):
     ::
         # Uniform distribution with 5 points on [-2,2]
         >> Parameter(points=5, lower=-2, upper=2, param_type='Uniform')
-
         # Gaussian distribution with 3 points with mean=4.0, variance=2.5
         >> Parameter(points=3, shape_parameter_A=4, shape_parameter_B=2.5, 
         param_type='Gaussian') 
-
         # Gamma distribution with 15 points with k=1.0 and theta=2.0
         >> Parameter(points=15, shape_parameter_A=1.0, shape_parameter_B=2.0, 
         param_type='Gamma')
-
         # Exponential distribution with 12 points with lambda=0.5
         >> Parameter(points=12, shape_parameter_A=0.5, param_type='Exponential')
-
     """
 
     # constructor
@@ -92,11 +87,9 @@ class Parameter(object):
     def computeMean(self):
         """
         Returns the mean of the parameter 
-
         :param Parameter self: An instance of the Parameter class
         :return: mu, mean of the parameter
         :rtype: double
-
         **Sample declaration**
         :: 
             >> var1 = Parameter(points=12, shape_parameter_A=0.5, param_type='Exponential')
@@ -115,62 +108,45 @@ class Parameter(object):
         return mu
 
 
-    def getPDF(self, N, graph=None):
+    def getPDF(self, N):
         """
-        Returns the PDF of the parameter 
-
+        Returns the probability density function of the parameter 
         :param Parameter self: An instance of the Parameter class
         :param integer N: Number of points along the x-axis 
-        :param boolean graph: Select 1 for python to plot the PDF on a graph. Default is 0, in which case
-            no graph is produced 
         :return: x, 1-by-N matrix that contains the values of the x-axis along the support of the parameter 
         :rtype: ndarray
         :return: w, 1-by-N matrix that contains the values of the PDF of the parameter
         :rtype: ndarray
-
-        Raises:
-            AttributeError: The ``Raises`` section is a list of all exceptions
-                that are relevant to the interface.
-            ValueError: If `param2` is equal to `param1`.
-
         **Sample declaration**
         :: 
             >> var1 = Parameter(points=12, shape_parameter_A=0.5, param_type='Exponential')
             >> x, y = var1.getPDF(50)
         """
         if self.param_type is "Gaussian":
-            x, y = analytical.Gaussian(N, self.shape_parameter_A, self.shape_parameter_B)
+            x, y = analytical.PDF_GaussianDistribution(N, self.shape_parameter_A, self.shape_parameter_B)
         elif self.param_type is "Beta":
-            x, y = analytical.BetaDistribution(N, self.shape_parameter_A, self.shape_parameter_B, self.lower, self.upper) 
+            x, y = analytical.PDF_BetaDistribution(N, self.shape_parameter_A, self.shape_parameter_B, self.lower, self.upper) 
         elif self.param_type is "Gamma":
-            x, y = analytical.Gamma(N, self.shape_parameter_A, self.shape_parameter_B)
+            x, y = analytical.PDF_Gamma(N, self.shape_parameter_A, self.shape_parameter_B)
         elif self.param_type is "Weibull":
-            x, y = analytical.WeibullDistribution(N, self.shape_parameter_A, self.shape_parameter_B)
[…]
