+++ effective_quadratures/analyticaldistributions.py
@@ -1,12 +1,9 @@
 #!/usr/bin/env python
 """Sample probability density functions"""
 import numpy as np
-from scipy.special import erf, erfinv, gamma, beta
-#-----------------------------------------------------------------------
-#
-#    Analytical definitions of iCDFs
-#    To do: inverse Beta CDF
-#-----------------------------------------------------------------------
+from scipy.special import erf, erfinv, gamma, beta, betainc, gammainc
+
+# inverse-CDFs
 def iCDF_Gaussian(xx, mu, sigma):
     return mu + sigma * np.sqrt(2.0) * erfinv(2.0*xx - 1.0)
 
@@ -20,7 +17,6 @@ def iCDF_ExponentialDistribution(xx, lambda_value):
     return (-np.log(1.0 - xx))/(lambda_value)
 
 def iCDF_BetaDistribution(xx, a, b, lower, upper):
-    # Insert code for computing an inverse beta CDF here!
     # Insert code for computing an inverse beta CDF here!
      [x,y]=BetaDistribution(1000,a,b,lower,upper)
      c=[]
@@ -44,49 +40,100 @@ def iCDF_TruncatedGaussian(xx, mu, sigma, a, b):
             yy[i,0] = 0
     return yy
 
-#-----------------------------------------------------------------------
-#
-#    Analytical definitions of PDFs
-#
-#-----------------------------------------------------------------------
-def UniformDistribution(N, lower, upper):
+# Cumulative distribution
+def CDF_GaussianDistribution(N, mu, sigma):
+    x = np.linspace(-15*sigma, 15*sigma, N)
+    x = x + mu # scaling it by the mean!
+    w = 0.5*(1 + erf((x - mu)/(sigma * np.sqrt(2) ) ) )
+    return x, w
+
+def CDF_TruncatedGaussianDistribution(N, mu, sigma, a, b):
+    def cumulative(x):
+        return 0.5 * (1 + erf(x/np.sqrt(2)))
+    x = np.linspace(a, b, N) 
+    zeta = (x - mu)/(sigma)
+    alpha = (a - mu)/(sigma)
+    beta = (b - mu)/(sigma)
+    Z = cumulative(beta) - cumulative(alpha)
+    w = (cumulative(zeta) - cumulative(alpha))/(Z)
+    return x, w
+
+def CDF_BetaDistribution(N, a, b, lower, upper):
+    x = np.linspace(0, 1, N)
+    w = np.zeros((N,1))
+    for i in range(0, N):
+        w[i] = betainc(a, b, x[i])
+    return x, w
+
+def CDF_WeibullDistribution(N, lambda_value, k):
+    x = np.linspace(0, 15/k, N)
+    w = 1 - np.exp(-1.0 * ( (x) / (lambda_value * 1.0)  )**k )
+    return x, w
+
+def CDF_UniformDistribution(N, lower, upper):
+     x = np.linspace(lower, upper, N)
+     w = np.zeros((N, 1))
+     for i in range(0, N):
+         w[i] = (x[i] - lower)/(upper - lower)
+     return x, w
+
+def CDF_GammaDistribution(N, k, theta):
+    x = np.linspace(0, k*theta*10, N)
+    w = 1.0/(gamma(k)) * gammainc(k, x/theta)
+    return x, w
+
+def CDF_CauchyDistribution(N, x0, gammavalue):
+    x = np.linspace(-15*gammavalue, 15*gammavalue, N)
+    x = x + x0
[…]
