### 2017-02-18 65156be
more updates -- travis still failing.
