+++ examples/Integration.py
@@ -1,0 +1,13 @@
+"How to compute Integrals in effective quadratures!"
+
+
+# 1D example
+numint = Integration([10], 'Gauss-Legendre')
+p, w = numint
+
+
+
+# To do:
+# 1. Finish integration class and test function!
+# 2. CLean up "where it fails?"
+# 3. Clean up gradients!
