+++ effective_quadratures/base.py
@@ -5,6 +5,7 @@ from indexset import IndexSet
 from effectivequads import EffectiveSubsampling
 from computestats import Statistics
 import integrals as integrals
+import plotting as plotting
 import analyticaldistributions as analytical
 from utils import error_function, evalfunction, meshgrid
 from qr import qr_MGS, solveLSQ
