### 2017-02-22 3e62be2
updates to plotting -- tentative!
