+++ effective_quadratures/plotting.py
@@ -42,7 +42,7 @@ def coeffplot2D(coefficients, index_set, x_label, y_label, filename=None):
     G = np.log10(np.abs(z))
     Zm = np.ma.masked_where(np.isnan(G),G)
     opacity = 0.8
-    plt.rc('text', usetex=True)
+    #plt.rc('text', usetex=True)
     plt.rc('font', family='serif')
     mpl.rcParams['axes.linewidth'] = 2.0
     fig = plt.figure()
@@ -73,7 +73,7 @@ def coeffplot2D(coefficients, index_set, x_label, y_label, filename=None):
 def bestfit(x_train, y_train, x_test, y_test, CI, x_label, y_label, filename=None):
 
     opacity = 0.8
-    plt.rc('text', usetex=True)
+    #plt.rc('text', usetex=True)
     plt.rc('font', family='serif')
     mpl.rcParams['axes.linewidth'] = 2.0
     fig = plt.figure()
@@ -116,7 +116,7 @@ def bestfit3D(x_train, y_train, x_test, y_test, x_label, y_label, z_label, filen
     xx1 = x_test[0]
     xx2 = x_test[1]
     opacity = 0.8
-    plt.rc('text', usetex=True)
+    #plt.rc('text', usetex=True)
     plt.rc('font', family='serif')
     mpl.rcParams['axes.linewidth'] = 2.0
     mpl.rc('axes', edgecolor='white', labelcolor='black', grid=True)
@@ -152,7 +152,7 @@ def bestfit3D(x_train, y_train, x_test, y_test, x_label, y_label, z_label, filen
 
 def parameterplot(x_axis, y_pdf, y_cdf, filename=None, x_label=None, y_label1=None, y_label2=None):
     opacity = 0.8
-    plt.rc('text', usetex=True)
+    #plt.rc('text', usetex=True)
     plt.rc('font', family='serif')
     mpl.rcParams['axes.linewidth'] = 2.0
 
@@ -194,7 +194,7 @@ def parameterplot(x_axis, y_pdf, y_cdf, filename=None, x_label=None, y_label1=No
 
 def lineplot(x, y, x_label, y_label, filename=None):
     opacity = 0.8
-    plt.rc('text', usetex=True)
+    #plt.rc('text', usetex=True)
     plt.rc('font', family='serif')
     mpl.rcParams['axes.linewidth'] = 2.0
     fig = plt.figure()
@@ -224,7 +224,7 @@ def scatterplot3D(x, f, x1_label=None, x2_label=None, f_label=None, filename=Non
         raise(ValueError, 'scatterplot(x, y): The number of rows in x must be equivalent to the number of rows in y')
     
     opacity = 0.8
-    plt.rc('text', usetex=True)
+    #plt.rc('text', usetex=True)
     plt.rc('font', family='serif')
     mpl.rcParams['axes.linewidth'] = 2.0
     mpl.rc('axes', edgecolor='white', labelcolor='black', grid=True)
@@ -261,25 +261,30 @@ def scatterplot3D(x, f, x1_label=None, x2_label=None, f_label=None, filename=Non
     else:
         plt.show()
 
-def scatterplot(x, y, x_label, y_label, filename=None):
+def scatterplot(x, y, x_label, y_label, filename=None, marker_type=None, color_choice=None):
     x = np.mat(x)
     y = np.mat(y)
     m, n = x.shape
     p, q = y.shape
-    if n > m :
-        raise(ValueError, 'scatterplot(x, y): Matrix x of size m-by-n, must satisfy m>=n')
-    if m != p:
-        raise(ValueError, 'scatterplot(x, y): The number of rows in x must be equivalent to the number of rows in y')
+    if marker_type is None:
[…]
