### 2017-03-20 a3dbc3a
added approximation utility in eq.
