+++ effective_quadratures/effectivequads.py
@@ -93,12 +93,49 @@ class EffectiveSubsampling(object):
         self.row_indices = None
         self.dimensions = len(uq_parameters)
     
+    def integrate(function):
+        """
+        Add integration utility here!
+
+        """
+        # Determine the index set to be used!
+        dimensions = len(stackOfParameters)
+        orders = []
+        flags = []
+        uniform = 1
+        not_uniform = 0
+        for i in range(0, dimensions):
+            orders.append(int(stackOfParameters[i].order - 1) )
+            if stackOfParameters[i].param_type is 'Uniform':
+                flags.append(uniform)
+            else:
+                flags.append(not_uniform)
+
+        # Define the hyperbolic cross
+        hyperbolic = IndexSet('Hyperbolic basis', orders=orders, q=q_parameter)
+        maximum_number_of_evals = hyperbolic.getCardinality()
+        effectiveQuads = EffectiveSubsampling(stackOfParameters, hyperbolic)
+
+        # Call the effective subsampling object
+        points = effectiveQuads.getEffectivelySubsampledPoints(maximum_number_of_evals)
+        xn = effectiveQuads.solveLeastSquares(maximum_number_of_evals, function)
+        integral_esq = xn[0]
+
+    # For normalizing!
+    for i in range(0, dimensions):
+        if flags[i] == 0:
+            integral_esq  = integral_esq
+        elif flags[i] == 1:
+            integral_esq = integral_esq * (stackOfParameters[i].upper - stackOfParameters[i].lower )
+    
+    return integral_esq[0], points
+    
     def set_no_of_evals(self, no_of_evals):
         """
         Sets the number of model evaluations the user wishes to afford for generating the polynomial. 
 
         :param EffectiveSubsampling object: An instance of the EffectiveSubsampling class.
-        :param integer no_of_evals: The number of subsampled the user requires
+        :param integer no_of_evals: The number of subsamples the user requires
 
         **Sample usage:** 
         For useage please see the ipython-notebooks at www.effective-quadratures.org
@@ -153,15 +190,11 @@ class EffectiveSubsampling(object):
     
     def least_no_of_subsamples_reqd(self):
         """
-        Returns the coefficients for the effectively subsampled quadratures least squares problem. 
+        Computes the least number of subsamples required when using effectively subsampled quadratures. 
 
         :param EffectiveSubsampling object: An instance of the EffectiveSubsampling class
-        :param integer maximum_number_of_evals: The maximum number of evaluations the user would like. This value has to be atleast equivalent to the
-            total number of basis terms of the index set.    
-        :param callable function_values: A function call to the simulation model, that takes in d inputs and returns one output. If users know the 
-            quadrature subsamples required, they may also input all the simulation outputs as a single ndarray.     
-        :return: x, the coefficients of the least squares problem.
-        :rtype: ndarray
+        :return: points, The least number of subsamples required. In the absence of gradients, this function simply returns the number of basis terms. In the presence of gradients this function uses an iterative rank-determination algorithm to compute the number of subsamples required.
+        :rtype: int
 
         **Sample declaration**
         :: 
@@ -186,7 +219,7 @@ class EffectiveSubsampling(object):
         :param EffectiveSubsampling object: An instance of the EffectiveSubsampling class.
         :param callable function_values: A callable function or a numpy matrix of model evaluations at the quadrature subsamples.
         :param callable gradient_values: A callable function of a numpy matrix of gradient evaluations at the quadrature subsamples.
-        :param string technique: The least squares technique to be used; options include: 'weighted' (default), 'constrainedDE', 'constrainedNS'. These options only matter when using gradient evaluations. They correspond to a stacked / weighted least squares approach, a constrained approach using       direct elimination, and a constrained approach using the null space method. 
[…]
