+++ effective_quadratures/analyticaldistributions.py
@@ -133,7 +133,7 @@ def PDF_ExponentialDistribution(N, lambda_value):
     w = lambda_value * np.exp(-lambda_value * x)
     return x, w
 
-def  PDF_TruncatedGaussianDistribution(N, mu, sigma, a, b):
+def  PDF_TruncatedGaussian(N, mu, sigma, a, b):
     x = np.linspace(a, b, N)
     w = 1.0/( np.sqrt(2 * sigma**2 * np.pi)) * np.exp(-(x - mu)**2 * 1.0/(2 * sigma**2) )
     w = 1.0/sigma * w
@@ -141,3 +141,7 @@ def  PDF_TruncatedGaussianDistribution(N, mu, sigma, a, b):
     second_term = GaussianCDF(a, mu, sigma)
     w = w / (first_term - second_term)
     return x, w
+
+def GaussianCDF(constant, mu, sigma):
+    w = 1.0/2 * (1 + erf((constant - mu)/(sigma * np.sqrt(2))) )
+    return w
