+++ WIP/spence.py
@@ -1,181 +1,0 @@
-import numpy as np
-import scipy.special as sp
-import copy as cp
-
-def fn_orpolyval(x,n,an,bn,orpoly):
-    if n == 0:
-        p1 = (x-an[0])*orpoly[:,0]
-    else:
-        p1 = (x-an[n])*orpoly[:,n]-bn[n]*orpoly[:,n-1]
-    return p1
-    
-def fn_erfun_quad(a,b,h,tol):
-    abscissa = lambda k,h,a,b: 0.5*((b-a)*sp.erf(k*h)+(a+b))
-    weight = lambda k,h,a,b: (b-a)*np.exp(-(k*h)**2)/np.sqrt(np.pi)
-    k = 0
-    uk1 = np.array([abscissa(k,h,a,b)])
-    wgt = weight(k,h,a,b)
-    wk1 = np.array([wgt])
-    while wgt > tol:
-        k = k+1
-        uk1 = np.append(uk1,[abscissa(k,h,a,b),abscissa(-k,h,a,b)])
-        wgt = weight(k,h,a,b)
-        wk1 = np.append(wk1,[wgt,wgt])
-    
-    lgt = 2*k+1
-    m = np.arange(1,k+1,1)    
-    uk = np.zeros(lgt)
-    wk = np.zeros(lgt)
-    cz = k+1
-    uk[cz-1] = uk1[0]
-    wk[cz-1] = wk1[0]
-    ixp = cz+m
-    ixm = cz-m
-    ixp1 = 2*m
-    ixm1 = ixp1+1
-    uk[ixp-1] = uk1[ixp1-1]
-    uk[ixm-1] = uk1[ixm1-1]
-    wk[ixp-1] = wk1[ixp1-1]
-    wk[ixm-1] = wk1[ixm1-1]
-    return uk, wk
-    
-def fn_qrs_public(n,flag):
-    h = 0.004
-    tol = 1e-17
-    if flag==1: #QRS45_azimuthal, [0,pi/2], pi/4 symmetry
-        a = -1./np.sqrt(2)
-        b = -a
-        wfn = lambda u: 1./np.sqrt(1-u**2)
-        uk, wk = fn_erfun_quad(a,b,h,tol)
-        tuk = uk
-    elif flag==2: #QRS60_azimuthal, [0,pi/3], pi/6 symmetry
-        a = -1./2
-        b = -a
-        wfn = lambda u: 1./np.sqrt(1-u**2)
-        uk, wk = fn_erfun_quad(a,b,h,tol)
-        tuk = uk
-    elif flag==3: #QRS90_azimuthal, [0,pi/2], pi/2 symmetry
-        a = 0
-        b = np.pi/2
-        wfn = lambda u: 1
-        [uk,wk] = fn_erfun_quad(a,b,h,tol)
-        tuk = 0.5*np.sin(uk)
-    elif flag==4: #QRA90_azimuthal_2D_RZ
-        a = 0
-        b = 1
-        wfn = lambda u: 2./np.sqrt(2-u**2)
-        [uk,wk] = fn_erfun_quad(a,b,h,tol)
-        tuk = uk*np.sqrt(2-uk**2)
-    elif flag==5: #QRA45_azimuthal, [0,pi/2], pi/4 symmetry (sin(0.5*(beta-pi/4)))
-        a = -np.sqrt(1-1./np.sqrt(2))/np.sqrt(2)
-        b = -a
-        wfn = lambda u: 1./np.sqrt(1-u**2)
-        [uk,wk] = fn_erfun_quad(a,b,h,tol)
-        tuk = uk
[…]
