### 2017-03-24 ec0c518
edited tests
