### 2017-03-24 6e4761f
cleaned up tests
