+++ tests/test_parameter.py
@@ -16,7 +16,6 @@ class TestParameter(TestCase):
 
         # Parameter test 2: getRecurrenceCoefficients()
         var2 = Parameter(points=15, param_type='Uniform', lower=-1, upper=1)
-        x, y = var2.getPDF(300, graph=1)
         ab = var2.getRecurrenceCoefficients()
         print ab
         print '\n'
