+++ .travis.yml
@@ -4,6 +4,7 @@ sudo: false
 language: python
 python:
   - "2.7"
+  - "3.5"
 
 addons:
   apt:
