+++ effective_quadratures/polyreg.py
@@ -1,0 +1,194 @@
+"""Polynomial regression"""
+from parameter import Parameter
+from indexset import IndexSet
+import numpy as np
+from math import factorial
+from itertools import combinations
+from utils import error_function, evalfunction, find_repeated_elements, meshgrid
+from plotting import bestfit, bestfit3D, histogram
+from qr import solveLSQ
+from computestats import Statistics
+
+class Polyreg(object):
+    """
+    This class defines a polyfit object
+
+    :param training_x: A numpy 
+    :param IndexSet index_set: An instance of the IndexSet class, in case the user wants to overwrite the indices
+        that are obtained using the orders of the univariate parameters in Parameters uq_parameters. The latter 
+        corresponds to a tensor grid index set and is the default option if no index_set parameter input is given.
+    
+    **Sample declarations** 
+    ::
+        >> s = Parameter(lower=-2, upper=2, param_type='Uniform', points=4)
+        >> T = IndexSet('Total order', [3,3])
+        >> polyObject = Polynomial([s,s],T) # basis is defined by T
+
+        >> s = Parameter(lower=-2, upper=2, param_type='Uniform')
+        >> polyObject = Polynomial([s,s]) # Tensor basis is used
+    """
+    # Constructor
+    def __init__(self, training_x, training_y, option):
+        self.training_x = training_x
+        self.training_y = training_y
+        self.option = option
+        X = self.training_x
+        Y = self.training_y
+
+        m, n = X.shape
+        ones = np.ones((m, 1))
+        self.dimensions = n
+        #total_terms = nchoosek(n) + order, order)        
+
+        if self.option is 'linear':
+            A = np.mat(np.hstack([X, ones]), dtype='float64')
+            coeffs, not_used = solveLSQ(A, Y)
+            self.coefficients =  np.mat(coeffs, dtype='float64')
+            self.Xmatrix = A
+
+        elif self.option is 'quadratic':
+           dimensions = n
+           variables = range(0, dimensions)
+           combination = list(combinations(variables, 2)) 
+           constants = np.mat(np.ones((m, 1)), dtype='float64')
+
+            # Compute the interaction terms!
+           XC = np.mat( np.ones((m, len(combination))) , dtype='float64')
+           for i in range(0, len(combination) ):
+                for j in range(0, m):
+                    XC[j,i] = X[j, combination[i][0] ] * X[j, combination[i][1] ] ; 
+
+           # Compute the squared terms
+           X2 = np.mat(np.ones((m, dimensions ) ) , dtype = 'float64')
+           for i in range(0, dimensions ):
+                for j in range(0, m):
+                    X2[j,i] = X[j, i] * X[j,i ] ; 
+
+           # Set up the A matrix
+           A = np.mat(np.hstack([constants, X, X2, XC]), dtype = 'float64' )
+           self.Xmatrix = A
+           coeffs, not_used = solveLSQ(A, Y)
+           self.coefficients = np.mat(coeffs, dtype='float64')
+        else:
+            raise(ValueError, 'PolyFit.__init__: invalid fitting option: Choose between linear or quadratic.')
+
+    # Test Polynomial
+    def testPolynomial(self, test_x):
+        """
+        Returns the PDF of the model output. This routine effectively multiplies the coefficients of a polynomial
+        expansion with its corresponding basis polynomials. 
+    
+        :param PolyFit self: An instance of the PolyFit class
+        :param: numpy-matrix test_x: The function that needs to be approximated (or interpolated)
+        :return: polyapprox: The polynomial expansion of a function
+        :rtype: numpy matrix
+
+        """
+        coefficients = self.coefficients
+        p, q = test_x.shape
[…]
