+++ effective_quadratures/integrals.py
@@ -1,273 +1,0 @@
-#!/usr/bin/env python
-"""Utilities for computing integrals of functions"""
-from parameter import Parameter
-from polynomial import Polynomial
-from indexset import IndexSet
-from effectivequads import EffectiveSubsampling
-import numpy as np
-#****************************************************************************
-# Functions to code:
-#    
-# 1. Faster Gauss quadrature rules (see work by Alex Townsend & Nick Hale)
-# 2. Sparse grid quadrature rules with different growth rules
-# 3. Spherical quadrature rules
-# 4. Padua quadrature rules -- add 2D!
-#****************************************************************************
-"""
-class EQ_Integration(object):
-   
-    Quadrature rules for integrating a function over the interval [-1,1]^d.
-    For multivariate quadrature rules on non-tensorial domains, this routine will use effective subsampling
-    to compute appropriate points and weights.
-
-    :param int order: Maximum order along each direction
-    :param basis: For multivariate domains, user may select 'Total order', 'Hyperbolic cross' or 'Euclidean'. 
-        The hyperbolic cross parameter is set by default to 0.5.
-    :param method: Determines the type of quadrature rule to be used. Available options are: 
-        Gauss-Legendre or Clenshaw-Curtis.
-    :param coords: The domain over which the integral is to be computed. Options are, 'Cartesian' (default) or 'Spherical'
-
-    **Sample declarations** 
-    ::
-        # Uniform distribution with 5 points on [-2,2]
-        >> Parameter(points=5, lower=-2, upper=2, param_type='Uniform')
-
-    
-    def __init__(self, order, basis=None, method=None, coords=None):
-        
-        self.order = order
-        dimension = len(order)
-
-        if  
-        self.basis = basis
-
-        if basis is None
-        p, w = 
-        self.points = 
-
-
-"""
-def getCartesianPointsAndWeights(order, method):
-    if method is 'Gauss':
-        p, w = getGauss(order)
-    elif method is 'Clenshaw-Curtis':
-        p, w = getClenshawCurtis(order)
-    elif method is 'Padua':
-        p, w = getPadua(order)
-    return 0
-         
-    
-def tensorgrid(stackOfParameters, function=None):
-    """
-    Computes a tensor grid of quadrature points based on the distributions for each Parameter in stackOfParameters 
-
-    :param Parameter array stackOfParameters: A list of Parameter objects
-    :param callable function: The function whose integral needs to be computed. Can also be input as an array of function values at the
-        quadrature points. If the function is given as a callable, then this routine outputs the integral of the function and an array of
-        the points at which the function was evaluated at to estimate the integral. These are the quadrature points. In case the function is
-        not given as a callable (or an array, for that matter), then this function outputs the quadrature points and weights. 
-      
-    :return: tensor_int: The tensor grid approximation of the integral
-    :rtype: double
-    :return: points:  The quadrature points
-    :rtype: numpy ndarray
-    :return: weights: The quadrature weights
-    :rtype: numpy ndarray
-
-    **Notes**
-    For further details on this routine, see Polynomial.getPointsAndWeights()
-
-    """
-    # Determine the index set to be used!
-    dimensions = len(stackOfParameters)
-    orders = []
-    flags = []
-    uniform = 1
-    not_uniform = 0
-    for i in range(0, dimensions):
-        orders.append(stackOfParameters[i].order)
-        if stackOfParameters[i].param_type is 'Uniform':
-            flags.append(uniform)
-        else:
-            flags.append(not_uniform)
-
-    tensor = IndexSet('Tensor grid', orders)
-    polyObject = Polynomial(stackOfParameters, tensor)
-
[…]
