+++ effective_quadratures/__init__.py
@@ -1,15 +1,9 @@
 'Initialization file'
-import effective_quadratures.parameter, effective_quadratures.polynomial, effective_quadratures.indexset, effective_quadratures.effectivequads, effective_quadratures.computestats, effective_quadratures.integrals, effective_quadratures.plotting, effective_quadratures.analyticaldistributions, effective_quadratures.utils, effective_quadratures.qr
-#from effective_quadratures.base import *
-
-# Set default logging handler to avoid "No handler found" warnings.
-#import logging
-#try:  # Python 2.7+
-#    from logging import NullHandler
-#except ImportError:
-#    class NullHandler(logging.Handler):
-#        def emit(self, record):
-#            pass
-
-# remove stderr logger
-#logging.getLogger(__name__).addHandler(NullHandler())
+from parameter import Parameter
+from polyint import Polyint
+from polylsq import Polylsq
+from polyreg import Polyreg
+from stats import Statistics
+import qr as qr 
+from indexset import IndexSet
+from utils import evalfunction, evalgradients
