+++ effective_quadratures/base.py
@@ -1,11 +1,0 @@
-import numpy as np
-from effective_quadratures.parameter import Parameter
-from effective_quadratures.polynomial import Polynomial
-from effective_quadratures.indexset import IndexSet
-from effective_quadratures.effectivequads import EffectiveSubsampling
-from effective_quadratures.computestats import Statistics
-import effective_quadratures.integrals as integrals
-import effective_quadratures.plotting as plotting
-import effective_quadratures.analyticaldistributions as analytical
-from effective_quadratures.utils import error_function, evalfunction, meshgrid
-from effective_quadratures.qr import qr_MGS, solveLSQ

[…]
