+++ equadratures/utils.py
@@ -132,8 +132,3 @@ def find_repeated_elements(index_value, matrix):
             if( all(matrix[j,1::] == selected_cell_indices)  ): # If all the indices match---i.e., the specific index is repeated
                 local_store = np.append(local_store, [j])
     return local_store
-
-# An error function
-def error_function(string_value):
-    print string_value
-    sys.exit()
