+++ equadratures/__init__.py
@@ -1,11 +1,9 @@
 'Initialization file'
 from polyint import Polyint
 from parameter import Parameter
-"""
 from polylsq import Polylsq
 from polyreg import Polyreg
 from stats import Statistics
 import qr as qr 
 from indexset import IndexSet
-from utils import evalfunction, evalgradients
-"""
+from utils import evalfunction, evalgradients, meshgrid
