+++ tests/test_polylsq.py
@@ -1,34 +1,30 @@
 #!/usr/bin/env python
 from unittest import TestCase
 import unittest
-from effective_quadratures.polynomial import PolyFit
-from effective_quadratures.utils import meshgrid
+from equadratures import *
 import numpy as np
-from mpl_toolkits.mplot3d import Axes3D
-import matplotlib.pyplot as plt
-from matplotlib import cm
 
-class TestPolyFit(TestCase):
+class TestPolylsq(TestCase):
+    
+    def fun(x):
+        return 1.0/(1 + 50*(x[0]- 0.9)**2 + 50*(x[1] + 0.9)**2 )
+    
+    
+    no_of_subsamples = 10
+    x1 = Parameter(param_type="Uniform", lower=-1, upper=1, points=no_of_subsamples)
+    x2 = Parameter(param_type="Uniform", lower=-1, upper=1, points=no_of_subsamples)
+    parameters = [x1, x2]
+    Hyperbolic = IndexSet("Hyperbolic basis", orders=[no_of_subsamples-1,no_of_subsamples-1], q=0.3)
+    esq = Polylsq(parameters, Hyperbolic)
+    minimum_subsamples = esq.least_no_of_subsamples_reqd() 
+    esq.set_no_of_evals(minimum_subsamples)
+    psmall = esq.subsampled_quadrature_points
+    xvec = np.linspace(-1.,1.,40)                               
+    x,y = np.meshgrid(xvec, xvec)
+    z =  1.0/(1 + 50*(x - 0.9)**2 + 50*(y + 0.9)**2 )       
+    stackOfPoints, x1, x2 = meshgrid(-1.0, 1.0, 40, 40)
+    zapprox = esq.getPolynomialApproximation(stackOfPoints, fun)
 
-    def test_linear(self):
-
-        x_train = np.mat([[0.0, 32.1205],
-        [0.0714, 29.7601],
-        [0.1429, 30.6798],
-        [0.2023, 22.1313],
-        [0.2857, 27.6573],
-        [0.3571, 26.5302],
-        [0.4001, 18.63],
-        [0.4286, 25.0],
-        [0.5714, 23.1148],
-        [0.6429, 21.2932],
-        [0.7143, 22.4207],
-        [0.7857, 21.1507],
-        [0.8310, 22.91923],
-        [0.9286, 19.5951],
-        [1.0000, 18.022]], dtype='float64')
-        y_train = np.mat([6.8053,-1.5184,1.6416,2.0121,6.3543,14.3442,25.3121,16.4426,18.1953,28.9913,27.2246,40.3759,45.12322, 55.3726,72.0], dtype='float64')
-        poly1 = PolyFit(x_train, y_train.T, 'quadratic')
 
 if __name__ == '__main__':
     unittest.main()
