+++ equadratures/polyint.py
@@ -1,4 +1,4 @@
-"""Operations involving multivariate polynomials (without gradients)"""
+"""Operations involving multivariate polynomials (without gradients) via numerical quadrature"""
 from parameter import Parameter
 from indexset import IndexSet
 import numpy as np
