+++ docs/_sources/stats.txt
@@ -1,0 +1,5 @@
+Statistics
+===========
+
+.. automodule:: equadratures.stats
+    :members:
