+++ README.md
@@ -1,11 +1,22 @@
 ![EFFECTIVE-QUADRATURES](https://static.wixstatic.com/media/dad873_3938470ea83849db8b53716c94dd20e8~mv2.png/v1/fill/w_269,h_66,al_c,usm_0.66_1.00_0.01/dad873_3938470ea83849db8b53716c94dd20e8~mv2.png)
 
 # Effective Quadratures
-[![Gitter](https://badges.gitter.im/gitterHQ/gitter.svg)](https://gitter.im/Effective-Quadratures)  [![Build Status](https://travis-ci.org/Effective-Quadratures/Effective-Quadratures.svg?branch=dev)](https://travis-ci.org/Effective-Quadratures/Effective-Quadratures)
 [![MIT licensed](https://img.shields.io/badge/license-MIT-blue.svg)](https://raw.githubusercontent.com/Effective-Quadratures/Effective-Quadratures/main/LICENSE.rst)
-[![Coverage Status](https://coveralls.io/repos/github/Effective-Quadratures/Effective-Quadratures/badge.svg?branch=dev-new)](https://coveralls.io/github/Effective-Quadratures/Effective-Quadratures?branch=dev-new)
 
-Suite of tools for generating polynomials for approximation, uncertainty quantification, optimization and integration. For further details, quick-start guides and papers see:
+**What is Effective Quadratures?**
+Effective Quadratures is a suite of tools for generating polynomials for approximation, uncertainty quantification (UQ), optimization and integration.  
+
+**Why do we need it?**
+  To replicate and compare state-of-the-art techniques and results in UQ;
+  To facilitate easy programming of new techniques at a research level;
+  To have a unified, easy-to-use platform with interactive demos---tailored for industrial use.
+
+**So, what's new in Effective Quadratures?**
+  New sampling & integration routines for more efficienct use of your model evaluations;
+  Techniques that leverage adjoint (gradient) information;
+  Plotting subroutines for quick, one-command plots.
+
+For further details, notebooks and papers see:
 <br>
 www.effective-quadratures.org
 <br>
@@ -15,14 +26,27 @@ For installation on Mac and Linux systems, simply type the following into the te
 ```bash
 > sudo python setup.py install
 ```
-For installation on Windows, you will need [Anaconda] (https://www.continuum.io/downloads#windows); select the Python 2.7 option. Upon successful installation, open the Sypder IDE and go into the Effective-Quadratures-master directory and type the following in the command window
+For installation on Windows, you will need [Anaconda](https://www.continuum.io/downloads#windows); select the Python 2.7 option. Upon successful installation, open the Sypder IDE and go into the Effective-Quadratures-master directory and type the following in the command window
 ```bash
 > python setup.py install
 ```
-This should build the code. Just make sure you include the location of effective_quadratures folder to your python file and you should be good to go.
+This should build the code. Just make sure you include the location of effective_quadratures folder to your python file and you should be good to go. To run this code you will require python 2.7, numpy, scipy and matplotlib. 
+
+# Simple example of use
+Below we provide two simple examples that showcase some of the functionality in Effective Quadratures. 
+```python
+from equadratures import *
+
+x = Parameter(param_type='Gaussian', shape_parameter_A=3.0, shape_parameter_B=2.0, points=6)
+y = Parameter(param_type='Weibull', shape_parameter_A=1.0, shape_parameter_B=2.2, points=4)
+
+p = Polyint([x,y])
+points, weights = p.getPointsAndWeights()
+plotting.scatterplot(points[:,0], points[:,1], '$x$', '$y$', marker_type='o', color_choice='crimson')
+```
 
 # Documentation
[…]
