### 2017-03-26 0d8b7b8
Update paper.md
