### 2017-03-26 #166 @psesh
Hi @katyhuff, apologizes for the delay in getting back. I have addressed the reviewer's comments. Here is a summary of what I've done:

1. Added references & a statement of need to the software paper;
2. Added community guidelines, a simple example, & a statement of need to the README.md file;
3. Added plenty of examples (which can be found at: www.effective-quadratures.org). The README.md file has a link to the website. I found it easier to put all the examples in the website instead of the sphinx generated html files.
4. Updated (and cleaned up) the code to make it easier to use. The current version is 5.2 (which is what I am submitting)

Do let me know if this is alright! Thanks again :)
