### 2017-03-27 #166 @nicoguaro
@katyhuff, it seems like it is good to go. I would say that the "Community guidelines" might be a little bit more explicit. But, you let me know.
