### 2017-03-27 #166 @katyhuff
@psesh thank you for your updated paper. I believe this is ready to accept.

@nicoguaro Thanks for the speedy responses! It looks good to me, too! I agree, we certainly would encourage more detailed community guidelines, but I think the statement in the readme, combined with some additional encouragement for contribution on the website fits the bill.

@arfon, can you take it from here? This paper is ready to accept. Please note the version is now 5.2 and not 3 as stated in submission.
