### 2017-03-29 #166 @arfon
@nicoguaro many thanks for reviewing this and @katyhuff for editing ✨

@psesh - your paper is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00166 ⚡️ 🚀 💥
